refactor(gda-balancing): deepen structured operation seams (#645)
