+++ libs/gda-balancing/src/gda_balancing/domain/authority/admission.py
@@ -12,14 +12,6 @@ from typing import Any, cast
 import jsonschema
 
 from gda_balancing.domain.canonical import JsonValue, canonical_bytes, content_identity
-from gda_balancing.domain.structured_values import (
-    StructuredValueIndex,
-    StructuredValueFault,
-    equal_result_contract,
-    is_empty_result_contract,
-    language_structured_value_index,
-    lookup_type_contract,
-)
 from gda_balancing.domain.authority.graph import (
     LanguageBundleGraph,
     LanguageBundleIndex,
@@ -37,8 +29,7 @@ from gda_balancing.domain.authority.package_validation import (
 from gda_balancing.domain.authority.runtime_validation import (
     _operation_result_source_shape_is_closed,
     _runtime_authority_is_closed,
-    literal_operation_contracts,
-    operation_value_contract_matches,
+    derive_operation_value_contracts,
 )
 from gda_balancing.domain.authority.template_validation import (
     _template_admission_profiles_are_closed,
@@ -59,6 +50,7 @@ from gda_balancing.domain.authority.vector_validation import (
     _rule_is_closed,
     _vector_header_is_closed,
 )
+from gda_balancing.domain.operation_program import project_operation_program
 
 SCHEMA2_REFUSAL_STAGES = (
     "ingress",
@@ -3179,132 +3171,6 @@ def _literal_typing_profiles_are_closed(
     return True
 
 
-def _literal_matches_operation_contract(
-    value: Any,
-    formal: dict[str, Any],
-    literal_profiles: Any,
-    typed_envelope_contract: Any,
-    fixed_value_contracts: Any,
-) -> bool:
-    return len(
-        matches := [
-            contract
-            for contract in literal_operation_contracts(
-                value,
-                literal_profiles,
-                typed_envelope_contract,
-                fixed_value_contracts,
-            )
-            if operation_value_contract_matches(contract, formal)
-        ]
-    ) == 1 and bool(matches)
-
-
-def _operation_contract_for_structured_type(
-    type_expression: Any,
-    literal_profiles: Any,
-) -> dict[str, Any] | None:
-    if not isinstance(type_expression, dict):
-        return None
-    package = type_expression.get("package")
-    version = type_expression.get("version")
-    type_id = type_expression.get("id")
-    if all(isinstance(item, str) and item for item in (package, version, type_id)):
-        exact_type = {"id": type_id, "package": package, "version": version}
-        scalar_profiles = (
-            [
-                profile
-                for profile in literal_profiles
-                if isinstance(profile, dict)
-                and profile.get("source_kind") == "integer"
-                and profile.get("type") == exact_type
-            ]
-            if isinstance(literal_profiles, list)
-            else []
-        )
-        if len(scalar_profiles) == 1:
-            profile = scalar_profiles[0]
-            return {
-                member: profile[member]
-                for member in (
-                    "domain",
-                    "kind",
-                    "numeric_policy",
-                    "representation",
-                    "unit",
-                )
-            } | {
-                "type": exact_type,
-            }
-        return {"type": exact_type, "value_kind": "nominal-structured"}
-    if type_expression.get("kind") in {"list", "ref"}:
-        return {"type": type_expression, "value_kind": "nominal-structured"}
-    return None
-
-
-def _declared_lookup_operation_contract(
-    value_contract: dict[str, Any],
-    key: Any,
-    structured_authority: StructuredValueIndex,
-    literal_profiles: Any,
-    key_candidates: tuple[dict[str, Any], ...] | None,
-    runtime_numeric_policies: list[str],
-) -> tuple[dict[str, Any], str] | None:
-    type_expression = value_contract.get("type")
-    try:
-        selector, result_type, refusal_signal = lookup_type_contract(
-            type_expression, key, authority=structured_authority
-        )
-    except StructuredValueFault:
-        return None
-    if selector == "local-index":
-        integer_profiles = (
-            [
-                profile
-                for profile in literal_profiles
-                if isinstance(profile, dict)
-                and profile.get("source_kind") == "integer"
-                and profile.get("numeric_policy") in runtime_numeric_policies
-            ]
-            if isinstance(literal_profiles, list)
-            else []
-        )
-        if not key_candidates or any(
-            not any(
-                operation_value_contract_matches(candidate, profile)
-                for profile in integer_profiles
-            )
-            for candidate in key_candidates
-        ):
-            return None
-    result = _operation_contract_for_structured_type(result_type, literal_profiles)
-    return (result, refusal_signal) if result is not None else None
-
-
-def _declared_equal_result_contract(
-    value_contract: dict[str, Any],
-    structured_authority: StructuredValueIndex,
-) -> str | None:
-    try:
-        return equal_result_contract(
-            value_contract.get("type"), authority=structured_authority
-        )
-    except StructuredValueFault:
-        return None
-
-
-def _declared_is_empty_result_contract(
-    value_contract: dict[str, Any],
-    structured_authority: StructuredValueIndex,
-) -> str | None:
-    try:
-        return is_empty_result_contract(
-            value_contract.get("type"), authority=structured_authority
-        )
-    except StructuredValueFault:
-        return None
-
-
 def _operation_alias_policy_is_closed(operation: dict[str, Any]) -> bool:
     inputs = operation.get("inputs")
     policy = operation.get("alias_policy")
@@ -3364,8 +3230,6 @@ def _operation_composition_diagnostic_subjects(
     language = language_bundle.get("language")
     if not isinstance(language, dict) or not isinstance(language.get("packages"), list):
         return ("language.operations",)
-    literal_profiles = language.get("literal_typing_profiles")
-    literal_contract = kernel.get("meta_format", {}).get("literal_typing")
     invocation_contract = (
         kernel.get("meta_format", {})
         .get("runtime_program", {})
@@ -3375,34 +3239,20 @@ def _operation_composition_diagnostic_subjects(
     runtime_nodes = (
         runtime_program.get("nodes") if isinstance(runtime_program, dict) else None
     )
-    fixed_value_contracts = (
-        runtime_program.get("fixed_value_contracts")
-        if isinstance(runtime_program, dict)
-        else None
-    )
-    runtime_numeric_policies = (
-        runtime_program.get("numeric", {}).get("compatible_value_numeric_policies")
-        if isinstance(runtime_program, dict)
-        and isinstance(runtime_program.get("numeric"), dict)
-        else None
-    )
+    value_contracts = derive_operation_value_contracts(kernel, language_bundle)
     result_source_shapes = (
         invocation_contract.get("result_source_shapes")
         if isinstance(invocation_contract, dict)
         else None
     )
     if (
-        not isinstance(literal_contract, dict)
-        or literal_contract.get("selection") != "unique-formal-match"
-        or not isinstance(literal_profiles, list)
-        or not isinstance(result_source_shapes, dict)
+        not isinstance(result_source_shapes, dict)
         or not isinstance(runtime_nodes, list)
-        or not isinstance(fixed_value_contracts, dict)
-        or not isinstance(runtime_numeric_policies, list)
-        or not runtime_numeric_policies
-        or not all(isinstance(policy, str) for policy in runtime_numeric_policies)
+        or value_contracts is None
     ):
         return ("language.literal-typing-profiles",)
+    fixed_value_contracts = value_contracts.fixed_value_contracts
+    runtime_numeric_policies = value_contracts.runtime_numeric_policies
     node_definitions = {
         node["id"]: node
         for node in runtime_nodes
@@ -3410,12 +3260,6 @@ def _operation_composition_diagnostic_subjects(
     }
     if len(node_definitions) != len(runtime_nodes):
         return ("kernel.meta-format.runtime-program.nodes",)
-    try:
-        structured_authority = language_structured_value_index(
-            language_bundle, kernel=kernel
-        )
-    except (KeyError, TypeError, ValueError):
-        return ("language.structured-operations",)
     reasons_by_signal: dict[str, list[str]] = {}
     for reason in cast(list[dict[str, Any]], language.get("reasons", [])):
         signal = reason.get("signal") if isinstance(reason, dict) else None
@@ -3456,7 +3300,17 @@ def _operation_composition_diagnostic_subjects(
         for _owner, operation in operations.values()
     ):
         return ("language.operations.alias-policy",)
-    cache: dict[tuple[str, str, str], tuple[set[str], set[str], int]] = {}
+    operation_node_ids = {
+        node_id
+        for node_id, node in node_definitions.items()
+        if node["semantics"]["operator"] in {"invoke-operation", "schedule-operation"}
+    }
+    invocation_node_ids = {
+        node_id
+        for node_id, node in node_definitions.items()
+        if node["semantics"]["operator"] == "invoke-operation"
+    }
+    cache: dict[tuple[str, str, str], tuple[frozenset[str], frozenset[str], int]] = {}
     guard_body_keys: set[tuple[str, str, str]] = set()
     found: set[str] = set()
 
@@ -3466,7 +3320,7 @@ def _operation_composition_diagnostic_subjects(
     def close(
         key: tuple[str, str, str],
         stack: tuple[tuple[str, str, str], ...],
-    ) -> tuple[set[str], set[str], int] | None:
+    ) -> tuple[frozenset[str], frozenset[str], int] | None:
         if key in stack:
             owner, operation = operations[key]
             refuse(owner, operation, "cycle", "operation")
@@ -3504,7 +3358,6 @@ def _operation_composition_diagnostic_subjects(
         local_producers: dict[str, int] = {}
         effects = set(cast(list[str], operation["effects"]))
         refusals = set(cast(list[str], operation["refusals"]))
-        charge = 0
         seen_sites: set[str] = set()
         operation_result_sites: set[str] = set()
         source_producer_reached = False
@@ -3519,7 +3372,7 @@ def _operation_composition_diagnostic_subjects(
                 for candidate in candidate_sets[0]
                 if all(
                     any(
-                        operation_value_contract_matches(candidate, other)
+                        value_contracts.matches(candidate, other)
                         for other in candidates
                     )
                     for candidates in candidate_sets[1:]
@@ -3554,7 +3407,6 @@ def _operation_composition_diagnostic_subjects(
         for instruction_index, instruction in enumerate(
             cast(list[dict[str, Any]], operation["body"])
         ):
-            charge += 1
             node = node_definitions.get(instruction.get("node"))
             if not isinstance(node, dict) or set(instruction) != set(
                 node["required_members"]
@@ -3594,9 +3446,8 @@ def _operation_composition_diagnostic_subjects(
                                 not in {"nominal-structured", "structured"}
                                 or (
                                     isinstance(expected_result_contract, str)
-                                    and _declared_equal_result_contract(
-                                        candidate,
-                                        structured_authority,
+                                    and value_contracts.declared_equal_result_contract(
+                                        candidate
                                     )
                                     == expected_result_contract
                                 )
@@ -3616,10 +3467,7 @@ def _operation_composition_diagnostic_subjects(
                                     candidate
                                     for candidate in candidates
                                     if any(
-                                        operation_value_contract_matches(
-                                            candidate,
-                                            common,
-                                        )
+                                        value_contracts.matches(candidate, common)
                                         for common in shared
                                     )
                                 ),
@@ -3634,10 +3482,7 @@ def _operation_composition_diagnostic_subjects(
                             narrowed = tuple(
                                 candidate
                                 for candidate in candidates
-                                if operation_value_contract_matches(
-                                    candidate,
-                                    expected,
-                                )
+                                if value_contracts.matches(candidate, expected)
                             )
                             if not narrowed:
                                 refuse(
@@ -3713,7 +3558,7 @@ def _operation_composition_diagnostic_subjects(
                             target_variant["kind"] == "port"
                             and (
                                 target_value not in parent_ports
-                                or not operation_value_contract_matches(
+                                or not value_contracts.matches(
                                     parent_ports[target_value],
                                     fixed_value_contracts[
                                         target_variant["value_contract"]
@@ -3833,7 +3678,7 @@ def _operation_composition_diagnostic_subjects(
                         cache.pop(guard_key, None)
                     if guard_closure is None:
                         return None
-                    body_effects, body_refusals, body_charge = guard_closure
+                    body_effects, body_refusals, _body_charge = guard_closure
                     if not body_effects <= effects:
                         refuse(
                             owner,
@@ -3850,7 +3695,6 @@ def _operation_composition_diagnostic_subjects(
                             "refusals",
                         )
                         return None
-                    charge += body_charge
                 result_definition = cast(dict[str, Any], node["result"])
                 if result_definition["kind"] in {"local", "draw"}:
                     if (
@@ -3878,9 +3722,8 @@ def _operation_composition_diagnostic_subjects(
                                 (expected_contract,)
                                 if value_candidates
                                 and all(
-                                    _declared_is_empty_result_contract(
-                                        candidate,
-                                        structured_authority,
+                                    value_contracts.declared_is_empty_result_contract(
+                                        candidate
                                     )
                                     == typing["contract"]
                                     for candidate in value_candidates
@@ -3911,15 +3754,12 @@ def _operation_composition_diagnostic_subjects(
                                 result
                                 for candidate in value_candidates
                                 if (
-                                    result := _declared_lookup_operation_contract(
+                                    result := value_contracts.declared_lookup_contract(
                                         candidate,
                                         instruction.get("key"),
-                                        structured_authority,
-                                        literal_profiles,
                                         locals_.get(cast(str, instruction.get("key")))
                                         if isinstance(instruction.get("key"), str)
                                         else None,
-                                        runtime_numeric_policies,
                                     )
                                 )
                                 is not None
@@ -3945,12 +3785,7 @@ def _operation_composition_diagnostic_subjects(
                         )
                     else:
                         literal_candidates = [
-                            literal_operation_contracts(
-                                instruction.get(member),
-                                literal_profiles,
-                                literal_contract["typed_envelope_profile"],
-                                fixed_value_contracts,
-                            )
+                            value_contracts.literal_contracts(instruction.get(member))
                             for member in cast(list[str], typing["members"])
                         ]
                         result_candidates = compatible_candidates(literal_candidates)
@@ -4004,7 +3839,7 @@ def _operation_composition_diagnostic_subjects(
                     )
                     if (
                         actual is None
-                        or not operation_value_contract_matches(actual, formal)
+                        or not value_contracts.matches(actual, formal)
                         or (
                             formal["access"] in {"read-write", "write"}
                             and actual["access"] not in {"read-write", "write"}
@@ -4027,7 +3862,7 @@ def _operation_composition_diagnostic_subjects(
                             [
                                 actual
                                 for actual in actual_candidates
-                                if operation_value_contract_matches(actual, formal)
+                                if value_contracts.matches(actual, formal)
                             ]
                         )
                         != 1
@@ -4039,13 +3874,7 @@ def _operation_composition_diagnostic_subjects(
                     literal = operand.get("literal")
                     if formal[
                         "access"
-                    ] != "read" or not _literal_matches_operation_contract(
-                        literal,
-                        formal,
-                        literal_profiles,
-                        literal_contract["typed_envelope_profile"],
-                        fixed_value_contracts,
-                    ):
+                    ] != "read" or not value_contracts.literal_matches(literal, formal):
                         refuse(owner, operation, site, "arguments")
                         return None
                     alias_key = f"literal:{operand['literal']}"
@@ -4076,7 +3905,7 @@ def _operation_composition_diagnostic_subjects(
                 lexical_environment[name] = (child_result,)
                 local_producers[name] = 1
             elif result.get("kind") == "operation-result":
-                if not operation_value_contract_matches(
+                if not value_contracts.matches(
                     cast(dict[str, Any], child["result"]),
                     cast(dict[str, Any], operation["result"]),
                 ):
@@ -4154,16 +3983,13 @@ def _operation_composition_diagnostic_subjects(
             child_closure = close(cast(tuple[str, str, str], child_key), (*stack, key))
             if child_closure is None:
                 return None
-            child_effects, child_refusals, child_charge = child_closure
+            child_effects, child_refusals, _child_charge = child_closure
             if not child_effects <= set(cast(list[str], operation["effects"])):
                 refuse(owner, operation, site, "effects")
                 return None
             if not child_refusals <= set(cast(list[str], operation["refusals"])):
                 refuse(owner, operation, site, "refusals")
                 return None
-            effects.update(child_effects)
-            refusals.update(child_refusals)
-            charge += child_charge
 
         result_contract = cast(dict[str, Any], operation["result"])
         local_result_candidates = (
@@ -4180,7 +4006,7 @@ def _operation_composition_diagnostic_subjects(
             or (
                 source_kind == "port"
                 and isinstance(parent_ports.get(source.get("name")), dict)
-                and operation_value_contract_matches(
+                and value_contracts.matches(
                     cast(dict[str, Any], parent_ports[source["name"]]),
                     result_contract,
                 )
@@ -4197,7 +4023,7 @@ def _operation_composition_diagnostic_subjects(
                             tuple[dict[str, Any], ...],
                             local_result_candidates,
                         )
-                        if operation_value_contract_matches(
+                        if value_contracts.matches(
                             candidate,
                             result_contract,
                         )
@@ -4207,22 +4033,39 @@ def _operation_composition_diagnostic_subjects(
             )
             or (
                 source_kind == "unit"
-                and result_contract.get("type")
-                == {"package": "kernel", "version": "2.0.0", "id": "Unit"}
-                and result_contract.get("representation") == "Unit"
-                and result_contract.get("kind") == "unit"
-                and result_contract.get("unit") == "1"
-                and result_contract.get("domain") == {"kind": "unit"}
-                and result_contract.get("numeric_policy") == "exact-unit"
+                and value_contracts.matches(
+                    result_contract,
+                    cast(dict[str, Any], fixed_value_contracts["kernel-unit"]),
+                )
             )
         )
         if not source_is_compatible:
             found.add(f"language.operations.{owner}.{operation['id']}.result.source")
             return None
-        if charge > operation["resource_bounds"]["max_steps"]:
+        try:
+            projection = project_operation_program(
+                key,
+                {
+                    coordinate: definition
+                    for coordinate, (
+                        _definition_owner,
+                        definition,
+                    ) in operations.items()
+                },
+                operation_node_ids=operation_node_ids,
+                invocation_node_ids=invocation_node_ids,
+            )
+        except (KeyError, TypeError, ValueError):
+            refuse(owner, operation, "closure", "operation")
+            return None
+        if projection.resource_charge > operation["resource_bounds"]["max_steps"]:
             found.add(f"language.operations.{owner}.{operation['id']}.resource_bounds")
             return None
-        cache[key] = effects, refusals, charge
+        cache[key] = (
+            projection.effects,
+            projection.refusals,
+            projection.resource_charge,
+        )
         return cache[key]
 
     for key in sorted(operations):
