+++ libs/gda-balancing/src/gda_balancing/domain/authority/package_validation.py
@@ -7,7 +7,7 @@ from gda_balancing.domain.authority.package_semantics import (
     package_runtime_semantic_closure,
 )
 from gda_balancing.domain.authority.runtime_validation import (
-    operation_value_is_admitted,
+    derive_operation_value_contracts,
 )
 from gda_balancing.domain.authority.contract_validation import (
     _exact_path_value,
@@ -21,7 +21,6 @@ from gda_balancing.domain.authority.vector_validation import (
     _signed_int64,
     _value_program_instruction_is_closed,
 )
-from gda_balancing.domain.structured_values import language_structured_value_index
 
 
 def _package_is_closed(
@@ -327,28 +326,7 @@ def _package_evidence_vectors_are_closed(
         if isinstance(ordering, list)
         else None
     )
-    language = language_bundle.get("language")
-    literal_typing = kernel.get("meta_format", {}).get("literal_typing")
-    literal_profiles = (
-        language.get("literal_typing_profiles") if isinstance(language, dict) else None
-    )
-    typed_envelope_contract = (
-        literal_typing.get("typed_envelope_profile")
-        if isinstance(literal_typing, dict)
-        else None
-    )
-    fixed_value_contracts = (
-        runtime_program_contract.get("fixed_value_contracts")
-        if isinstance(runtime_program_contract, dict)
-        else None
-    )
-    resource_limit = kernel.get("resources", {}).get("max_ldb_admission_work")
-    try:
-        structured_authority = language_structured_value_index(
-            language_bundle, kernel=kernel
-        )
-    except (KeyError, TypeError, ValueError):
-        return False
+    value_contracts = derive_operation_value_contracts(kernel, language_bundle)
     if (
         not _package_vector_contract_is_closed(contract)
         or not isinstance(candidate_encoding, dict)
@@ -363,12 +341,7 @@ def _package_evidence_vectors_are_closed(
         or not all(isinstance(phase, str) and phase for phase in phase_rank)
         or not isinstance(runtime_nodes, list)
         or not all(isinstance(node, dict) for node in runtime_nodes)
-        or not isinstance(literal_profiles, list)
-        or not isinstance(typed_envelope_contract, dict)
-        or not isinstance(fixed_value_contracts, dict)
-        or not isinstance(resource_limit, int)
-        or isinstance(resource_limit, bool)
-        or resource_limit < 1
+        or value_contracts is None
     ):
         return False
     phase_inventory = set(phase_rank)
@@ -605,15 +578,7 @@ def _package_evidence_vectors_are_closed(
                 and set(item) == {"name", "value"}
                 and isinstance(item.get("name"), str)
                 and item["name"]
-                and operation_value_is_admitted(
-                    item.get("value"),
-                    formal,
-                    literal_profiles=literal_profiles,
-                    typed_envelope_contract=typed_envelope_contract,
-                    fixed_value_contracts=fixed_value_contracts,
-                    structured_authority=structured_authority,
-                    resource_limit=resource_limit,
-                )
+                and value_contracts.admits(item.get("value"), formal)
                 for item, formal in zip(values, operation_inputs, strict=False)
             )
             or len(values) != len(operation_inputs)
@@ -629,15 +594,7 @@ def _package_evidence_vectors_are_closed(
                 isinstance(item, dict)
                 and set(item) == set(kind["state_value_members"])
                 and isinstance(item.get("name"), str)
-                and operation_value_is_admitted(
-                    item.get("value"),
-                    formal,
-                    literal_profiles=literal_profiles,
-                    typed_envelope_contract=typed_envelope_contract,
-                    fixed_value_contracts=fixed_value_contracts,
-                    structured_authority=structured_authority,
-                    resource_limit=resource_limit,
-                )
+                and value_contracts.admits(item.get("value"), formal)
                 for item, formal in zip(state_after, state_inputs, strict=False)
             )
             or len(state_after) != len(state_inputs)
@@ -691,14 +648,8 @@ def _package_evidence_vectors_are_closed(
                 set(result) != {"kind", "value"}
                 or result.get("kind") != "value"
                 or not isinstance(operation.get("result"), dict)
-                or not operation_value_is_admitted(
-                    result.get("value"),
-                    cast(dict[str, Any], operation["result"]),
-                    literal_profiles=literal_profiles,
-                    typed_envelope_contract=typed_envelope_contract,
-                    fixed_value_contracts=fixed_value_contracts,
-                    structured_authority=structured_authority,
-                    resource_limit=resource_limit,
+                or not value_contracts.admits(
+                    result.get("value"), cast(dict[str, Any], operation["result"])
                 )
             ):
                 return False
