+++ libs/gda-balancing/src/gda_balancing/domain/evidence.py
@@ -13,51 +13,47 @@ from gda_balancing.domain.artifacts import (
     verify_artifact,
 )
 from gda_balancing.domain.publication import PublicationMember
+from gda_balancing.domain.operation_program import (
+    instruction_evaluation_sites,
+)
 from gda_balancing.domain.runtime.scheduler import RuntimeScheduler
 from gda_balancing.domain.experiment import (
     CheckedExperiment,
 )
-from gda_balancing.domain.runtime.execution import (
+from gda_balancing.domain.evidence_replay import (
+    ReplayInitializationProgramFault as _InitializationProgramFault,
+    attempted_operation_charge as _attempted_operation_charge,
+    evaluate_initialization_programs as _evaluate_initialization_programs,
+    execute_value_instruction as _execute_value_instruction,
+    replay_event_evidence as _replay_event_evidence,
+)
+from gda_balancing.domain.runtime.projections import (
     RuntimeRefusalOutcome,
-    _InitializationProgramFault,
-    _NamedRng,
-    _admit_declared_value,
-    _artifact,
-    _committed_event_projection,
-    _empty_runtime_journal_identity,
-    _evaluate_initialization_programs,
-    _evaluator_manifest,
-    _event_catalog_record,
-    _execute_value_instruction,
-    _extend_runtime_journal_identity,
-    _formula_programs_reachable_from_entrypoints,
-    _guard_expanded_instruction_indices,
-    _instruction_evaluation_sites,
-    _named_value_rows,
-    _integer_compare,
-    _metric_definition_identity,
-    _observation_event_id,
-    _operation_formula_evaluation_record,
-    _operation_formula_slot,
-    _ordered_root_events,
-    _pending_event_projection,
-    _projected_runtime_identity,
-    _reproduction_receipt,
-    _resolved_display_names,
-    _resolved_runtime_profile,
-    _root_event_id,
-    _runtime_contract,
-    _runtime_journal_contract,
-    _runtime_nodes,
-    _scenario_transition_events,
-    _scheduled_event_id,
-    _scheduler_contract,
+    artifact as _artifact,
+    committed_event_projection as _committed_event_projection,
+    empty_runtime_journal_identity as _empty_runtime_journal_identity,
+    evaluator_manifest as _evaluator_manifest,
+    event_catalog_record as _event_catalog_record,
+    extend_runtime_journal_identity as _extend_runtime_journal_identity,
+    formula_programs_reachable_from_entrypoints as _formula_programs_reachable_from_entrypoints,
+    metric_definition_identity as _metric_definition_identity,
+    named_value_rows as _named_value_rows,
+    observation_event_id as _observation_event_id,
+    operation_formula_slot as _operation_formula_slot,
+    ordered_root_events as _ordered_root_events,
+    pending_event_projection as _pending_event_projection,
+    projected_runtime_identity as _projected_runtime_identity,
+    resolved_display_names as _resolved_display_names,
+    resolved_runtime_profile as _resolved_runtime_profile,
+    root_event_id as _root_event_id,
+    reproduction_receipt as _reproduction_receipt,
+    runtime_contract as _runtime_contract,
+    runtime_journal_contract as _runtime_journal_contract,
+    runtime_nodes as _runtime_nodes,
+    scenario_transition_events as _scenario_transition_events,
+    scheduled_event_id as _scheduled_event_id,
+    scheduler_contract as _scheduler_contract,
 )
-from gda_balancing.domain.structured_values import (
-    StructuredValueFault,
-    selected_structured_value_index,
-)
-
 
 _INVALID_FORMULA_EVIDENCE = object()
 
@@ -287,7 +283,7 @@ def _event_formula_evaluations_are_authoritative(
             or evaluation.get("context") != site["context"]
             or evaluation.get("frame_identity") != event.get("snapshot_before_identity")
             or executions.get(call_path) != operation_reference["id"]
-            or site_identity not in _instruction_evaluation_sites(operation).values()
+            or site_identity not in instruction_evaluation_sites(operation).values()
             or (call_path, cast(str, site_identity)) in seen
         ):
             return False
@@ -599,383 +595,18 @@ def _replayed_event_evidence(
         catalog_by_id=catalog_by_id,
         events_by_id=events_by_id,
     )
-    parent_operation_id = parent_event.get("operation")
-    if root_arguments is None or not isinstance(parent_operation_id, str):
+    if root_arguments is None:
         return None
-    operations = {
-        cast(str, row["definition"]["id"]): row["definition"]
-        for row in checked.rir["selected_semantics"]["operations"]
-    }
-    formula_bindings_by_site = {
-        cast(str, cast(dict[str, Any], binding["site"])["identity"]): binding
-        for binding in cast(list[dict[str, Any]], checked.rir["formula_bindings"])
-        if cast(dict[str, Any], binding["site"])["kind"] == "operation-slot"
-    }
-    root_operation = operations.get(parent_operation_id)
-    if root_operation is None:
-        return None
-    declarations = _resolved_declarations(checked)
-    display_names = _resolved_display_names(declarations)
-    state: dict[bytes, JsonValue] = {
-        identity: values_by_name[display_name]
-        for identity, display_name in display_names.items()
-        if display_name
-        in (
-            values_by_name := {
-                cast(str, row["name"]): row["value"]
-                for row in cast(
-                    list[dict[str, JsonValue]], parent_event["state_before"]
-                )
-            }
-        )
-    }
-    actual_values = root_arguments[2]
-    calls = cast(list[dict[str, JsonValue]], parent_event["calls"])
-    schedules = cast(list[dict[str, JsonValue]], parent_event["schedules"])
-    draws = cast(list[dict[str, JsonValue]], parent_event["rng_draws"])
-    draw_index = 0
-    formula_evaluations: list[dict[str, JsonValue]] = []
-    rng = _NamedRng(
-        cast(int, checked.value["seed"]["value"]),
-        cast(dict[str, Any], _runtime_contract(checked)["named_rng"]),
+    return _replay_event_evidence(
+        checked,
+        parent_event,
+        parent_spec,
+        target_schedule,
+        root_arguments,
+        scenario_id=scenario_id,
+        catalog_by_id=catalog_by_id,
+        events_by_id=events_by_id,
     )
-
-    def consume_authoritative_draw(
-        traced: dict[str, JsonValue],
-    ) -> int | None:
-        try:
-            value, index, candidate, accepted = rng.draw(
-                cast(str, traced["stream"]),
-                cast(int, traced["minimum"]),
-                cast(int, traced["maximum"]),
-            )
-        except (KeyError, TypeError, ValueError):
-            return None
-        expected = {
-            "stream": traced["stream"],
-            "index": index,
-            "candidate_hex": rng.encode_candidate(candidate),
-            "accepted": accepted,
-            "minimum": traced["minimum"],
-            "maximum": traced["maximum"],
-            "value": value,
-        }
-        return value if traced == expected else None
-
-    for prior_event in sorted(
-        events_by_id.values(), key=lambda row: cast(int, row["index"])
-    ):
-        if cast(int, prior_event["index"]) >= cast(int, parent_event["index"]):
-            break
-        prior_record = catalog_by_id.get(cast(str, prior_event["event_id"]))
-        if prior_record is None:
-            return None
-        if prior_record["scenario"] != scenario_id:
-            continue
-        if any(
-            consume_authoritative_draw(draw) is None
-            for draw in cast(list[dict[str, JsonValue]], prior_event["rng_draws"])
-        ):
-            return None
-    numeric = cast(dict[str, Any], _runtime_contract(checked)["numeric"])
-    node_contracts = _runtime_nodes(checked)
-    structured_authority = selected_structured_value_index(
-        cast(dict[str, Any], checked.rir["selected_semantics"]),
-        kernel=checked.kernel,
-    )
-    structured_resource_limit = cast(
-        int, checked.language_bundle["resources"]["max_rule_match_steps"]
-    )
-    schedule_identity = _scheduler_contract(checked)["call_site_identity"]["schedule"]
-
-    def execute(
-        operation: dict[str, Any],
-        arguments: dict[str, JsonValue],
-        state_references: dict[str, dict[str, JsonValue]],
-        call_path: tuple[str, ...],
-    ) -> tuple[
-        str,
-        JsonValue,
-        tuple[dict[str, JsonValue], dict[str, dict[str, JsonValue]]] | None,
-    ]:
-        nonlocal draw_index
-        operation_before = dict(state)
-        variables: dict[str, Any] = dict(arguments)
-        extensions = operation.get("extensions", {})
-        snapshot_operands = (
-            extensions.get("standard.snapshot-operands")
-            if isinstance(extensions, dict)
-            else None
-        )
-        if isinstance(snapshot_operands, dict):
-            for row in cast(
-                list[dict[str, Any]], snapshot_operands.get("operands", [])
-            ):
-                identity = canonical_bytes(cast(JsonValue, row["resolved_symbol"]))
-                if identity not in actual_values:
-                    return "", None, None
-                variables[cast(str, row["name"])] = actual_values[identity]
-        operation_results: dict[str, JsonValue] = {}
-        outcome = cast(str, operation["default_outcome"])
-        evaluation_sites = _instruction_evaluation_sites(operation)
-        for instruction_index, instruction in enumerate(
-            cast(list[dict[str, Any]], operation["body"])
-        ):
-            evaluation_site_identity = evaluation_sites.get(instruction_index)
-            node_contract = node_contracts[instruction["node"]]
-            operator = node_contract["semantics"]["operator"]
-            if operator == "invoke-operation":
-                child = operations.get(cast(str, instruction["operation"]["id"]))
-                if child is None:
-                    return "", None, None
-                child_arguments: dict[str, JsonValue] = {}
-                child_state_references: dict[str, dict[str, JsonValue]] = {}
-                for binding in instruction["arguments"]:
-                    operand = binding["operand"]
-                    name = cast(str, binding["port"])
-                    if operand["kind"] == "port":
-                        source = cast(str, operand["port"])
-                        child_arguments[name] = cast(JsonValue, variables[source])
-                        if source in state_references:
-                            child_state_references[name] = state_references[source]
-                    elif operand["kind"] == "local":
-                        child_arguments[name] = cast(
-                            JsonValue, variables[operand["local"]]
-                        )
-                    else:
-                        child_arguments[name] = cast(JsonValue, operand["literal"])
-                child_path = (*call_path, cast(str, instruction["site"]))
-                child_outcome, child_result, found = execute(
-                    child,
-                    child_arguments,
-                    child_state_references,
-                    child_path,
-                )
-                if found is not None:
-                    return "", None, found
-                call = next(
-                    (
-                        row
-                        for row in calls
-                        if row["site"] == "/".join(child_path)
-                        and cast(dict[str, JsonValue], row["operation"])["id"]
-                        == child["id"]
-                    ),
-                    None,
-                )
-                if (
-                    call is None
-                    or cast(dict[str, JsonValue], call["outcome"])["id"]
-                    != child_outcome
-                ):
-                    return "", None, None
-                result_binding = instruction["result"]
-                if result_binding["kind"] == "local":
-                    variables[result_binding["name"]] = child_result
-                elif result_binding["kind"] == "operation-result":
-                    operation_results[instruction["site"]] = child_result
-                for alias, target in state_references.items():
-                    variables[alias] = state[canonical_bytes(cast(JsonValue, target))]
-                mapping = next(
-                    row
-                    for row in instruction["outcomes"]
-                    if row["outcome"] == child_outcome
-                )
-                if mapping["action"]["kind"] == "propagate":
-                    outcome = cast(str, mapping["action"]["outcome"])
-                    break
-                continue
-            if operator == "schedule-operation":
-                child_arguments = {}
-                child_state_references = {}
-                for binding in instruction["arguments"]:
-                    operand = binding["operand"]
-                    name = cast(str, binding["port"])
-                    if operand["kind"] == "port":
-                        source = cast(str, operand["port"])
-                        child_arguments[name] = cast(JsonValue, variables[source])
-                        if source in state_references:
-                            child_state_references[name] = state_references[source]
-                    elif operand["kind"] == "local":
-                        child_arguments[name] = cast(
-                            JsonValue, variables[operand["local"]]
-                        )
-                    else:
-                        child_arguments[name] = cast(JsonValue, operand["literal"])
-                call_site_identity = content_identity(
-                    cast(str, schedule_identity["domain"]),
-                    cast(
-                        JsonValue,
-                        {
-                            "parent_event_id": parent_event["event_id"],
-                            "parent_operation": operation["id"],
-                            "site": instruction["site"],
-                            "operation": instruction["operation"],
-                        },
-                    ),
-                )
-                if (
-                    target_schedule is not None
-                    and target_schedule["call_site_identity"] == call_site_identity
-                    and target_schedule["call_path"] == "/".join(call_path)
-                ):
-                    return "", None, (child_arguments, child_state_references)
-                scheduled = next(
-                    (
-                        row
-                        for row in schedules
-                        if row["call_site_identity"] == call_site_identity
-                        and row["call_path"] == "/".join(call_path)
-                    ),
-                    None,
-                )
-                if scheduled is None:
-                    return "", None, None
-                variables[instruction["result"]["name"]] = scheduled["event_id"]
-                continue
-            if operator == "cancel-event":
-                continue
-            if operator == "gameplay-precondition":
-                if not _integer_compare(
-                    node_contract["semantics"]["comparison"],
-                    cast(int, variables[instruction["left"]]),
-                    cast(int, variables[instruction["right"]]),
-                ):
-                    outcome = cast(str, instruction["outcome"])
-                    break
-            elif operator == "typed-require":
-                if variables[instruction["condition"]] != instruction["expected"]:
-                    return "", None, None
-            elif operator == "guarded-outcome-block":
-                if variables[instruction["condition"]]:
-                    guarded_operation = {
-                        "body": instruction["body"],
-                        "default_outcome": instruction["outcome"],
-                        "extensions": {},
-                        "id": operation["id"],
-                        "outcomes": list(operation["outcomes"]),
-                        "result": {"source": {"kind": "unit"}},
-                    }
-                    guarded_outcome, _guarded_result, found = execute(
-                        guarded_operation,
-                        variables,
-                        state_references,
-                        call_path,
-                    )
-                    if found is not None:
-                        return "", None, found
-                    if guarded_outcome != instruction["outcome"]:
-                        return "", None, None
-                    outcome = cast(str, instruction["outcome"])
-                    break
-            elif operator == "named-integer-draw":
-                if draw_index >= len(draws):
-                    return "", None, None
-                draw = draws[draw_index]
-                draw_index += 1
-                if (
-                    draw["stream"] != instruction["stream"]
-                    or draw["minimum"] != instruction["minimum"]
-                    or draw["maximum"] != instruction["maximum"]
-                ):
-                    return "", None, None
-                value = consume_authoritative_draw(draw)
-                if value is None:
-                    return "", None, None
-                variables[instruction["target"]] = value
-            elif node_contract["family"] == "expression":
-                _execute_value_instruction(
-                    instruction,
-                    variables,
-                    numeric,
-                    node_contract,
-                    structured_authority=structured_authority,
-                    structured_resource_limit=structured_resource_limit,
-                )
-            elif operator in {"state-integer-subtract", "state-write"}:
-                formal = cast(str, instruction["symbol"])
-                target = canonical_bytes(cast(JsonValue, state_references[formal]))
-                declaration = declarations.get(target)
-                if declaration is None:
-                    return "", None, None
-                state[target] = _admit_declared_value(
-                    cast(int, state[target])
-                    - cast(int, variables[instruction["value"]])
-                    if operator == "state-integer-subtract"
-                    else variables[instruction["value"]],
-                    numeric,
-                    declaration,
-                    structured_authority=structured_authority,
-                    structured_resource_limit=structured_resource_limit,
-                )
-                for alias, alias_target in state_references.items():
-                    if canonical_bytes(cast(JsonValue, alias_target)) == target:
-                        variables[alias] = state[target]
-            else:
-                return "", None, None
-            if (
-                evaluation_site_identity is not None
-                and evaluation_sites.get(instruction_index + 1)
-                != evaluation_site_identity
-            ):
-                binding = formula_bindings_by_site.get(evaluation_site_identity)
-                if binding is None:
-                    return "", None, None
-                evaluation = _operation_formula_evaluation_record(
-                    operation,
-                    binding,
-                    variables,
-                    evaluation_site_identity=evaluation_site_identity,
-                    frame_identity=cast(
-                        JsonValue, parent_event["snapshot_before_identity"]
-                    ),
-                    call_path=call_path,
-                )
-                if evaluation is None:
-                    return "", None, None
-                formula_evaluations.append(evaluation)
-        outcome_definition = next(
-            row for row in operation["outcomes"] if row["id"] == outcome
-        )
-        if outcome_definition["state_policy"] == "rollback":
-            state.clear()
-            state.update(operation_before)
-        result_source = operation["result"]["source"]
-        if outcome_definition["kind"] != "success":
-            result: JsonValue = None
-        elif result_source["kind"] in {"local", "port"}:
-            result = cast(JsonValue, variables[result_source["name"]])
-        elif result_source["kind"] == "operation-result":
-            result = operation_results[result_source["site"]]
-        else:
-            result = None
-        return outcome, result, None
-
-    root_state_references = {name: target for name, target in root_arguments[1].items()}
-    root_path = (
-        (cast(str, cast(dict[str, JsonValue], parent_event["entrypoint"])["id"]),)
-        if parent_event.get("entrypoint") is not None
-        else (f"scheduled:{parent_spec['call_site_identity']}",)
-    )
-    try:
-        replayed_outcome, _result, found = execute(
-            root_operation,
-            root_arguments[0],
-            root_state_references,
-            root_path,
-        )
-    except (
-        KeyError,
-        OverflowError,
-        StopIteration,
-        StructuredValueFault,
-        TypeError,
-        ValueError,
-    ):
-        return None
-    if not replayed_outcome and found is None:
-        return None
-    return found, formula_evaluations
 
 
 def _replayed_schedule_arguments(
@@ -1757,291 +1388,6 @@ def _formula_charge_through_evaluation_site(
         if not progressed:
             return None
     return None
-
-
-def _attempted_operation_charge(
-    checked: CheckedExperiment,
-    refusing_event: dict[str, Any],
-    refusing_event_spec: dict[str, Any],
-    *,
-    node_steps_before_operation: int,
-    bounds: dict[str, int],
-    require_budget_breach: bool,
-) -> int | None:
-    evaluation_site_identity = refusing_event.get("evaluation_site_identity")
-    target_instruction_index = refusing_event.get("instruction_index")
-    target_path = refusing_event.get("call_path")
-    if (
-        not isinstance(target_path, str)
-        or not isinstance(target_instruction_index, int)
-        and not isinstance(evaluation_site_identity, str)
-    ):
-        return None
-    operations = {
-        cast(str, row["definition"]["id"]): row["definition"]
-        for row in checked.rir["selected_semantics"]["operations"]
-    }
-    if refusing_event_spec["kind"] == "transition-invocation":
-        entrypoint = next(
-            (
-                row
-                for row in checked.rir["entrypoints"]
-                if row["id"] == refusing_event_spec["entrypoint"]
-            ),
-            None,
-        )
-        root_operation_id = (
-            cast(str, entrypoint["operation"]["id"]) if entrypoint is not None else None
-        )
-    elif refusing_event_spec["kind"] == "scheduled-transition":
-        operation_reference = refusing_event_spec.get("operation")
-        root_operation_id = (
-            cast(str, operation_reference["id"])
-            if isinstance(operation_reference, dict)
-            and isinstance(operation_reference.get("id"), str)
-            else None
-        )
-    else:
-        return None
-    root_operation = operations.get(root_operation_id) if root_operation_id else None
-    if root_operation is None:
-        return None
-    root_path = target_path.split("/", 1)[0]
-    calls = cast(list[dict[str, JsonValue]], refusing_event["attempted_calls"])
-    used_calls: set[int] = set()
-    node_contracts = _runtime_nodes(checked)
-    event_charge = 0
-    node_steps = node_steps_before_operation
-
-    def evaluation_sites(operation: dict[str, Any]) -> dict[int, str]:
-        extensions = operation.get("extensions")
-        if not isinstance(extensions, dict):
-            return {}
-        provenance = extensions.get("standard.instruction-provenance")
-        if (
-            not isinstance(provenance, dict)
-            or provenance.get("kind") != "instruction-evaluation-sites"
-            or not isinstance(provenance.get("sites"), list)
-        ):
-            return {}
-        return {
-            cast(int, row["instruction_index"]): cast(
-                str, row["evaluation_site_identity"]
-            )
-            for row in cast(list[dict[str, Any]], provenance["sites"])
-        }
-
-    def is_target(
-        operation: dict[str, Any],
-        call_path: str,
-        instruction_index: int,
-        sites: dict[int, str],
-    ) -> bool:
-        return (
-            call_path == target_path
-            and operation["id"] == refusing_event["operation"]
-            and (
-                instruction_index == target_instruction_index
-                if isinstance(target_instruction_index, int)
-                else sites.get(instruction_index) == evaluation_site_identity
-            )
-        )
-
-    def charge_instruction(
-        operation: dict[str, Any],
-        operation_charge: int,
-        instruction: dict[str, Any],
-    ) -> tuple[int, bool]:
-        nonlocal event_charge, node_steps
-        amount = cast(
-            int, node_contracts[instruction["node"]]["resource_charge"]["amount"]
-        )
-        operation_charge += amount
-        event_charge += amount
-        node_steps += amount
-        breached = (
-            operation_charge > cast(int, operation["resource_bounds"]["max_steps"])
-            or event_charge > bounds["max_event_steps"]
-            or node_steps > bounds["max_node_steps"]
-        )
-        return operation_charge, breached
-
-    def completed_invocation(
-        instruction: dict[str, Any], parent_path: str
-    ) -> dict[str, Any] | None:
-        child_path = f"{parent_path}/{instruction['site']}"
-        child = operations.get(cast(str, instruction["operation"]["id"]))
-        call_rows = [
-            (index, row)
-            for index, row in enumerate(calls)
-            if row["site"] == child_path
-            and cast(dict[str, JsonValue], row["operation"])["id"]
-            == instruction["operation"]["id"]
-        ]
-        if child is None or len(call_rows) != 1:
-            return None
-        call_index, call = call_rows[0]
-        child_outcome = cast(str, cast(dict[str, JsonValue], call["outcome"])["id"])
-        if not completed_operation(child, child_path, child_outcome):
-            return None
-        mapping = next(
-            (row for row in instruction["outcomes"] if row["outcome"] == child_outcome),
-            None,
-        )
-        if mapping is None:
-            return None
-        used_calls.add(call_index)
-        return cast(dict[str, Any], mapping["action"])
-
-    def completed_operation(
-        operation: dict[str, Any],
-        call_path: str,
-        expected_outcome: str,
-    ) -> bool:
-        operation_charge = 0
-        outcome = cast(str, operation["default_outcome"])
-        sites = evaluation_sites(operation)
-        body = cast(list[dict[str, Any]], operation["body"])
-        expanded_indices = _guard_expanded_instruction_indices(body)
-        for body_index, instruction in enumerate(body):
-            instruction_index = expanded_indices[body_index]
-            operation_charge, breached = charge_instruction(
-                operation,
-                operation_charge,
-                instruction,
-            )
-            if breached or is_target(operation, call_path, instruction_index, sites):
-                return False
-            operator = node_contracts[instruction["node"]]["semantics"]["operator"]
-            if operator == "invoke-operation":
-                action = completed_invocation(instruction, call_path)
-                if action is None:
-                    return False
-                if action["kind"] == "propagate":
-                    outcome = cast(str, action["outcome"])
-                    break
-            elif operator == "guarded-outcome-block":
-                if instruction["outcome"] != expected_outcome:
-                    continue
-                for guard_instruction in cast(
-                    list[dict[str, Any]], instruction["body"]
-                ):
-                    operation_charge, breached = charge_instruction(
-                        operation,
-                        operation_charge,
-                        guard_instruction,
-                    )
-                    if breached:
-                        return False
-                    guard_operator = node_contracts[guard_instruction["node"]][
-                        "semantics"
-                    ]["operator"]
-                    if guard_operator != "invoke-operation":
-                        continue
-                    action = completed_invocation(guard_instruction, call_path)
-                    if action is None or action["kind"] == "propagate":
-                        return False
-                outcome = expected_outcome
-                break
-            elif (
-                operator == "gameplay-precondition"
-                and instruction["outcome"] == expected_outcome
-            ):
-                outcome = expected_outcome
-                break
-        return outcome == expected_outcome
-
-    def charge_to_target(
-        operation: dict[str, Any],
-        call_path: str,
-    ) -> bool:
-        operation_charge = 0
-        sites = evaluation_sites(operation)
-        body = cast(list[dict[str, Any]], operation["body"])
-        expanded_indices = _guard_expanded_instruction_indices(body)
-        for body_index, instruction in enumerate(body):
-            instruction_index = expanded_indices[body_index]
-            operation_charge, breached = charge_instruction(
-                operation,
-                operation_charge,
-                instruction,
-            )
-            target = is_target(operation, call_path, instruction_index, sites)
-            if breached or target:
-                return target and (breached or not require_budget_breach)
-            operator = node_contracts[instruction["node"]]["semantics"]["operator"]
-            if operator == "guarded-outcome-block":
-                guard_body = cast(list[dict[str, Any]], instruction["body"])
-                guard_start = instruction_index + 1
-                guard_stop = guard_start + len(guard_body)
-                target_is_directly_in_guard = (
-                    call_path == target_path
-                    and operation["id"] == refusing_event["operation"]
-                    and isinstance(target_instruction_index, int)
-                    and guard_start <= target_instruction_index < guard_stop
-                )
-                target_is_in_guard_call = any(
-                    node_contracts[guard_instruction["node"]]["semantics"]["operator"]
-                    == "invoke-operation"
-                    and (
-                        target_path == f"{call_path}/{guard_instruction['site']}"
-                        or target_path.startswith(
-                            f"{call_path}/{guard_instruction['site']}/"
-                        )
-                    )
-                    for guard_instruction in guard_body
-                )
-                target_is_in_guard = (
-                    target_is_directly_in_guard or target_is_in_guard_call
-                )
-                if not target_is_in_guard:
-                    continue
-                for guard_offset, guard_instruction in enumerate(guard_body):
-                    operation_charge, breached = charge_instruction(
-                        operation,
-                        operation_charge,
-                        guard_instruction,
-                    )
-                    guard_index = guard_start + guard_offset
-                    target = is_target(operation, call_path, guard_index, sites)
-                    if breached or target:
-                        return target and (breached or not require_budget_breach)
-                    guard_operator = node_contracts[guard_instruction["node"]][
-                        "semantics"
-                    ]["operator"]
-                    if guard_operator != "invoke-operation":
-                        continue
-                    child_path = f"{call_path}/{guard_instruction['site']}"
-                    child = operations.get(
-                        cast(str, guard_instruction["operation"]["id"])
-                    )
-                    if child is None:
-                        return False
-                    if target_path == child_path or target_path.startswith(
-                        f"{child_path}/"
-                    ):
-                        return charge_to_target(child, child_path)
-                    action = completed_invocation(guard_instruction, call_path)
-                    if action is None or action["kind"] == "propagate":
-                        return False
-                return False
-            if operator != "invoke-operation":
-                continue
-            child_path = f"{call_path}/{instruction['site']}"
-            child = operations.get(cast(str, instruction["operation"]["id"]))
-            if child is None:
-                return False
-            if target_path == child_path or target_path.startswith(f"{child_path}/"):
-                return charge_to_target(child, child_path)
-            action = completed_invocation(instruction, call_path)
-            if action is None or action["kind"] == "propagate":
-                return False
-        return False
-
-    reached_target = charge_to_target(root_operation, root_path)
-    if not reached_target or used_calls != set(range(len(calls))):
-        return None
-    return event_charge
 
 
 def _terminal_audit_is_valid(
