+++ libs/gda-balancing/src/gda_balancing/domain/experiment.py
@@ -30,6 +30,11 @@ from gda_balancing.infrastructure.input_bytes import (
     read_bounded_input_with_sha256,
 )
 from gda_balancing.domain.model import admit_resolved_model
+from gda_balancing.domain.operation_program import (
+    expanded_operation_body,
+    operation_coordinate,
+    selected_operation_index,
+)
 from gda_balancing.domain.publication import find_published_artifact
 from gda_balancing.domain.runtime.scheduler import RuntimeScheduler
 from gda_balancing.domain.structured_values import (
@@ -290,38 +295,6 @@ def _external_input_plan_is_admitted(
     return True
 
 
-def _expanded_operation_body(
-    operation: dict[str, Any],
-    operations: dict[str, dict[str, Any]],
-    visiting: frozenset[str] = frozenset(),
-) -> list[dict[str, Any]]:
-    """Expand generic Operation composition without package-specific dispatch."""
-    operation_id = cast(str, operation["id"])
-    if operation_id in visiting:
-        raise ValueError("admitted Operation composition is cyclic")
-    nested_visiting = visiting | {operation_id}
-
-    def expand(instructions: list[dict[str, Any]]) -> list[dict[str, Any]]:
-        expanded: list[dict[str, Any]] = []
-        for instruction in instructions:
-            expanded.append(instruction)
-            if instruction["node"] == "guard-block":
-                expanded.extend(expand(cast(list[dict[str, Any]], instruction["body"])))
-                continue
-            if instruction["node"] not in {"invoke", "schedule"}:
-                continue
-            operation_ref = cast(dict[str, Any], instruction["operation"])
-            invoked = operations.get(cast(str, operation_ref["id"]))
-            if invoked is None:
-                raise ValueError("admitted Operation composition target is absent")
-            expanded.extend(
-                _expanded_operation_body(invoked, operations, nested_visiting)
-            )
-        return expanded
-
-    return expand(cast(list[dict[str, Any]], operation["body"]))
-
-
 def derive_scenario_program_requirements(
     rir: dict[str, Any],
     entrypoint_id: str,
@@ -330,29 +303,29 @@ def derive_scenario_program_requirements(
 ) -> tuple[dict[str, list[str]], list[str]]:
     """Project one Scenario's evaluator contract from its admitted RIR."""
     selected = cast(dict[str, Any], rir["selected_semantics"])
-    operations = {
-        row["definition"]["id"]: row["definition"]
-        for row in cast(list[dict[str, Any]], selected["operations"])
-    }
+    operations = selected_operation_index(selected)
     entrypoints = {
         row["id"]: row for row in cast(list[dict[str, Any]], rir["entrypoints"])
     }
     entrypoint = entrypoints.get(entrypoint_id)
     if entrypoint is None:
         raise ValueError("Scenario entrypoint is absent from the selected RIR")
-    operation = operations.get(entrypoint["operation"]["id"])
+    root_coordinate = operation_coordinate(entrypoint["operation"])
+    operation = operations.get(root_coordinate)
     if operation is None:
         raise ValueError("Scenario Operation is absent from the selected RIR")
     if operation["runtime_profile"] != runtime_profile:
         raise ValueError("Scenario Operation requires another Runtime profile")
-    expanded_body = _expanded_operation_body(operation, operations)
+    expanded_body = expanded_operation_body(root_coordinate, operations)
     instruction_nodes = {instruction["node"] for instruction in expanded_body}
     requirements = {
         "operation_kinds": sorted(
             {
                 operation["operation_kind"],
                 *(
-                    operations[instruction["operation"]["id"]]["operation_kind"]
+                    operations[operation_coordinate(instruction["operation"])][
+                        "operation_kind"
+                    ]
                     for instruction in expanded_body
                     if instruction["node"] in {"invoke", "schedule"}
                 ),
