+++ libs/gda-balancing/tests/test_schema2_experiment_cli.py
@@ -1,5 +1,6 @@
 """Public RPG Experiment tracer for Standard Schema 2.0 (#540)."""
 
+import ast
 import hashlib
 import json
 import os
@@ -15,7 +16,11 @@ import gda_balancing.application.experiment_run as experiment_run_application_mo
 import gda_balancing.domain.experiment as experiment_admission_module
 import gda_balancing.domain.artifacts as artifacts_module
 import gda_balancing.domain.evidence as experiment_evidence_module
+import gda_balancing.domain.evidence_replay as evidence_replay_module
+import gda_balancing.domain.operation_program as operation_program_module
+import gda_balancing.domain.runtime.projections as runtime_projection_module
 import gda_balancing.domain.runtime.execution as experiment_runtime_module
+import schema2_operation_execution_conformance_support as operation_conformance_module
 import gda_balancing.interfaces.cli.experiment_check as experiment_check_command_module
 import gda_balancing.interfaces.cli.experiment_run as experiment_command_module
 import gda_balancing.domain.authority.context as authority_module
@@ -43,11 +48,20 @@ from schema2_scheduler_production_support import (
     scheduler_detector_inventory,
 )
 from schema2_operation_execution_production_support import (
+    _language_operation_index as production_operation_index,
     evaluate_operation_execution_vector,
 )
-from schema2_bootstrap_conformance_support import _consumer_b
+from schema2_operation_execution_independent_support import (
+    reference_execute_event,
+    reference_rng_draw,
+)
+from schema2_operation_execution_conformance_support import (
+    _operation_index as conformance_operation_index,
+    candidate_conformance_failures,
+    operation_execution_observations,
+    operation_execution_vectors,
+)
 from schema2_bootstrap_production_support import (
-    _consumer_a,
     _refresh_package_closure_and_reidentify,
 )
 from schema2_authority_support import mutable_authorities
@@ -533,797 +547,6 @@ def _member(receipt: dict[str, Any], logical_name: str) -> dict[str, Any]:
         if item["logical_name"] == logical_name
     )
     return json.loads(Path(locator).read_text(encoding="utf-8"))
-
-
-def _reference_compare(comparison: str, left: int, right: int) -> bool:
-    if comparison == "greater-than-or-equal":
-        return left >= right
-    if comparison == "less-than":
-        return left < right
-    if comparison == "less-than-or-equal":
-        return left <= right
-    raise AssertionError(f"unsupported comparison in authority: {comparison}")
-
-
-def _reference_rng_draw(
-    contract: dict[str, Any],
-    seed: int,
-    stream: str,
-    minimum: int,
-    maximum: int,
-    states: dict[str, int],
-    indices: dict[str, int],
-) -> dict[str, Any]:
-    mask = (1 << contract["word_bits"]) - 1
-    if stream not in states:
-        derivation = contract["stream_derivation"]
-        digest = hashlib.sha256(
-            stream.encode(contract["stream_name_encoding"])
-        ).digest()
-        digest_slice = derivation["digest_slice"]
-        start = digest_slice["offset"]
-        end = start + digest_slice["length"]
-        states[stream] = (
-            seed
-            + int.from_bytes(
-                digest[start:end],
-                derivation["byte_order"],
-            )
-        ) & mask
-        indices[stream] = 0
-    transition = contract["state_transition"]
-    state = (states[stream] + int(transition["increment_hex"], 16)) & mask
-    states[stream] = state
-    candidate = state
-    for step in transition["mix_steps"]:
-        candidate ^= candidate >> step["xor_shift_right"]
-        if "multiply_hex" in step:
-            candidate = (candidate * int(step["multiply_hex"], 16)) & mask
-    index = indices[stream]
-    indices[stream] = index + 1
-    value = minimum + candidate % (maximum - minimum + 1)
-    candidate_width = contract["candidate_encoding"]["width_bits"] // 4
-    return {
-        "stream": stream,
-        "index": index,
-        "candidate_hex": f"{candidate:0{candidate_width}x}",
-        "accepted": True,
-        "minimum": minimum,
-        "maximum": maximum,
-        "value": value,
-    }
-
-
-def _reference_fact_rows(values: dict[str, Any]) -> list[dict[str, Any]]:
-    rows = []
-    for name in sorted(values):
-        value = values[name]
-        if isinstance(value, bool):
-            rows.append({"name": name, "kind": "boolean", "boolean": value})
-        elif not isinstance(value, int):
-            rows.append({"name": name, "kind": "structured", "value": value})
-        else:
-            rows.append({"name": name, "kind": "integer", "integer": value})
-    return rows
-
-
-class _ReferenceRuntimeRefusal(Exception):
-    def __init__(self, reason: str):
-        super().__init__(reason)
-        self.reason = reason
-        self.operation: str | None = None
-        self.call_path: str | None = None
-        self.call_site_identity: str | None = None
-
-
-def _reference_execute_event(
-    kernel: dict[str, Any],
-    operation: dict[str, Any],
-    operations: dict[str, dict[str, Any]],
-    scenario: dict[str, Any],
-    *,
-    seed: int,
-    state_names: set[str] | None = None,
-    resolved_entrypoint: dict[str, Any] | None = None,
-    resolved_declarations: list[dict[str, Any]] | None = None,
-    resolved_call_sites: list[dict[str, Any]] | None = None,
-    resolved_initialization_programs: list[dict[str, Any]] | None = None,
-    language_bundle: dict[str, Any] | None = None,
-) -> dict[str, Any]:
-    runtime = kernel["meta_format"]["runtime_program"]
-    numeric = runtime["numeric"]
-    nodes = {row["id"]: row for row in runtime["nodes"]}
-    variables: dict[str | tuple[str, str, str], Any]
-    state_targets: set[str | tuple[str, str, str]]
-    display_names: dict[str | tuple[str, str, str], str]
-    if resolved_entrypoint is not None:
-        assert resolved_declarations is not None
-        declarations = {
-            (
-                row["resolved_symbol"]["model"],
-                row["resolved_symbol"]["module"],
-                row["resolved_symbol"]["name"],
-            ): row
-            for row in resolved_declarations
-        }
-        variables = {
-            (
-                row["target"]["model"],
-                row["target"]["module"],
-                row["target"]["name"],
-            ): row["value"]
-            for row in resolved_entrypoint["scenario_input_contract"]["initializers"]
-        }
-        variables.update(
-            {
-                (
-                    row["target"]["model"],
-                    row["target"]["module"],
-                    row["target"]["name"],
-                ): row["value"]
-                for row in scenario["assignments"]
-            }
-        )
-        pending_programs = list(resolved_initialization_programs or [])
-        reachable_formula_targets = {
-            (
-                operand["symbol"]["model"],
-                operand["symbol"]["module"],
-                operand["symbol"]["name"],
-            )
-            for binding in resolved_entrypoint["arguments"]
-            if (operand := binding["operand"])["kind"] == "symbol"
-        }
-        while True:
-            previous_target_count = len(reachable_formula_targets)
-            for program in pending_programs:
-                target = program["target"]
-                target_coordinate = (
-                    target["model"],
-                    target["module"],
-                    target["name"],
-                )
-                if target_coordinate not in reachable_formula_targets:
-                    continue
-                reachable_formula_targets.update(
-                    (
-                        operand["resolved_symbol"]["model"],
-                        operand["resolved_symbol"]["module"],
-                        operand["resolved_symbol"]["name"],
-                    )
-                    for row in program["inputs"]
-                    if (operand := row["operand"])["kind"] != "literal"
-                )
-            if len(reachable_formula_targets) == previous_target_count:
-                break
-        pending_programs = [
-            program
-            for program in pending_programs
-            if (
-                program["target"]["model"],
-                program["target"]["module"],
-                program["target"]["name"],
-            )
-            in reachable_formula_targets
-        ]
-        while pending_programs:
-            progressed = False
-            for program in list(pending_programs):
-                values: dict[str, int] = {}
-                ready = True
-                for row in program["inputs"]:
-                    operand = row["operand"]
-                    if operand["kind"] == "literal":
-                        values[row["name"]] = operand["value"]
-                        continue
-                    symbol = operand["resolved_symbol"]
-                    coordinate = (
-                        symbol["model"],
-                        symbol["module"],
-                        symbol["name"],
-                    )
-                    if coordinate not in variables:
-                        ready = False
-                        break
-                    values[row["name"]] = variables[coordinate]
-                if not ready:
-                    continue
-                for row in program["body"]:
-                    instruction = row["instruction"]
-                    node = instruction["node"]
-                    if node == "constant":
-                        result = instruction["literal"]
-                    elif node == "copy":
-                        result = values[instruction["value"]]
-                    elif node == "add":
-                        result = (
-                            values[instruction["left"]] + values[instruction["right"]]
-                        )
-                    elif node == "subtract":
-                        result = (
-                            values[instruction["left"]] - values[instruction["right"]]
-                        )
-                    elif node == "multiply":
-                        result = (
-                            values[instruction["left"]] * values[instruction["right"]]
-                        )
-                    elif node == "maximum":
-                        result = max(
-                            values[instruction["left"]],
-                            values[instruction["right"]],
-                        )
-                    else:
-                        assert node == "if"
-                        result = values[
-                            instruction[
-                                "when_true"
-                                if values[instruction["condition"]]
-                                else "when_false"
-                            ]
-                        ]
-                    assert numeric["minimum"] <= result <= numeric["maximum"]
-                    values[instruction["target"]] = result
-                target = program["target"]
-                result_source = program["result"]
-                variables[(target["model"], target["module"], target["name"])] = values[
-                    result_source["name"]
-                ]
-                pending_programs.remove(program)
-                progressed = True
-            assert progressed
-        state_targets = {
-            coordinate
-            for coordinate, declaration in declarations.items()
-            if declaration["role"] == "state"
-        }
-        display_names = {
-            coordinate: declaration["resolved_symbol"]["name"]
-            for coordinate, declaration in declarations.items()
-        }
-    else:
-        assert state_names is not None
-        legacy_variables: dict[str, Any] = {
-            row["name"]: row["value"] for row in scenario["values"]
-        }
-        variables = cast(
-            dict[str | tuple[str, str, str], Any],
-            legacy_variables,
-        )
-        state_targets = set(state_names)
-        display_names = {name: name for name in legacy_variables}
-    cells = {name: {"value": value} for name, value in variables.items()}
-    state_cells = {name: cells[name] for name in state_targets if name in cells}
-    declared_domains_by_cell = (
-        {
-            id(cells[coordinate]): declaration["domain"]
-            for coordinate, declaration in declarations.items()
-            if coordinate in state_cells
-            and declaration["domain_kind"] == "closed-interval"
-        }
-        if resolved_entrypoint is not None
-        else {}
-    )
-    before = {name: cell["value"] for name, cell in state_cells.items()}
-    rng_states: dict[str, int] = {}
-    rng_indices: dict[str, int] = {}
-    draws: list[dict[str, Any]] = []
-    calls: list[dict[str, Any]] = []
-    call_sites = {
-        (row["parent_operation"]["id"], row["site"]): row
-        for row in (resolved_call_sites or [])
-    }
-    language = (
-        language_bundle.get("language") if isinstance(language_bundle, dict) else None
-    )
-    nominal_types = {
-        (row["package"], row["version"], row["id"]): row
-        for row in (
-            language.get("nominal_types", []) if isinstance(language, dict) else []
-        )
-        if isinstance(row, dict)
-        and all(
-            isinstance(row.get(member), str) for member in ("package", "version", "id")
-        )
-    }
-    constructors = {
-        row["id"]: row
-        for row in (
-            language.get("constructors", []) if isinstance(language, dict) else []
-        )
-        if isinstance(row, dict) and isinstance(row.get("id"), str)
-    }
-    structured_operations = [
-        row
-        for row in (
-            language.get("structured_operations", [])
-            if isinstance(language, dict)
-            else []
-        )
-        if isinstance(row, dict)
-    ]
-
-    def structural(type_expression: Any) -> tuple[dict[str, Any], dict[str, Any]]:
-        definition = type_expression
-        constructor = None
-        if isinstance(type_expression, dict) and set(type_expression) >= {
-            "id",
-            "package",
-            "version",
-        }:
-            nominal = nominal_types[
-                (
-                    type_expression["package"],
-                    type_expression["version"],
-                    type_expression["id"],
-                )
-            ]
-            definition = nominal["definition"]
-            constructor = constructors[nominal["constructor"]]
-        if not isinstance(definition, dict):
-            raise AssertionError("structured type is not admitted")
-        if constructor is None:
-            matches = [
-                candidate
-                for candidate in constructors.values()
-                if candidate.get("value_rule", {}).get("definition_kind")
-                == definition.get("kind")
-            ]
-            assert len(matches) == 1
-            constructor = matches[0]
-        return definition, constructor
-
-    def structured_law(type_expression: Any, operator: str) -> dict[str, Any]:
-        _definition, constructor = structural(type_expression)
-        matches = [
-            row["law"]
-            for row in structured_operations
-            if row.get("owner_constructor") == constructor["id"]
-            and row.get("law", {}).get("operator") == operator
-        ]
-        assert len(matches) == 1
-        return matches[0]
-
-    def structured_lookup(envelope: Any, key: Any) -> dict[str, Any]:
-        assert isinstance(envelope, dict) and set(envelope) == {"type", "value"}
-        definition, constructor = structural(envelope["type"])
-        selected_law = structured_law(envelope["type"], "bounded-lookup")
-        rule = constructor["value_rule"]
-        if selected_law["selector"] == "static-field":
-            field = next(
-                row
-                for row in definition[rule["fields_member"]]
-                if row[rule["field_name_member"]] == key
-            )
-            result_type = field[rule["field_type_member"]]
-            result_value = envelope["value"][key]
-        else:
-            assert selected_law["selector"] == "local-index"
-            result_type = definition[rule["element_member"]]
-            result_value = envelope["value"][key]
-        if isinstance(result_type, dict) and result_type.get("kind") == "nominal":
-            result_type = {
-                member: result_type[member] for member in ("id", "package", "version")
-            }
-        return {"type": result_type, "value": result_value}
-
-    def structured_is_empty(envelope: Any) -> bool:
-        assert isinstance(envelope, dict) and set(envelope) == {"type", "value"}
-        definition, constructor = structural(envelope["type"])
-        selected_law = structured_law(envelope["type"], "collection-is-empty")
-        assert constructor["value_rule"]["operator"] == "bounded-list"
-        assert selected_law["result_contract"] == "kernel-boolean"
-        assert definition["kind"] == "list"
-        return not envelope["value"]
-
-    def exact(value: int, target: dict[str, Any] | None = None) -> int:
-        if not numeric["minimum"] <= value <= numeric["maximum"]:
-            raise _ReferenceRuntimeRefusal("runtime.numeric_overflow")
-        domain = (
-            declared_domains_by_cell.get(id(target)) if target is not None else None
-        )
-        if domain is not None and not domain["minimum"] <= value <= domain["maximum"]:
-            raise _ReferenceRuntimeRefusal("runtime.numeric_overflow")
-        return value
-
-    def execute(
-        selected: dict[str, Any],
-        arguments: dict[str, dict[str, Any]],
-        stack: tuple[str, ...] = (),
-        path: tuple[str, ...] = (),
-    ) -> tuple[str, Any]:
-        assert selected["id"] not in stack
-        locals_: dict[str, dict[str, Any]] = {}
-        operation_results: dict[str, Any] = {}
-        frame_cells = {id(cell): cell for cell in arguments.values()}
-        snapshot = {key: cell["value"] for key, cell in frame_cells.items()}
-        outcome = selected["default_outcome"]
-
-        def cell(name: str) -> dict[str, Any]:
-            if name in locals_:
-                return locals_[name]
-            return arguments[name]
-
-        def write_local(name: str, value: Any) -> None:
-            locals_[name] = {"value": value}
-
-        def project_integer(value: Any) -> int | None:
-            if isinstance(value, int) and not isinstance(value, bool):
-                return value
-            if (
-                isinstance(value, dict)
-                and isinstance(value.get("value"), int)
-                and not isinstance(value["value"], bool)
-            ):
-                return value["value"]
-            return None
-
-        def integer(value: Any) -> int:
-            projected = project_integer(value)
-            assert projected is not None
-            return projected
-
-        try:
-            for instruction in selected["body"]:
-                node = nodes[instruction["node"]]
-                assert set(instruction) == set(node["required_members"])
-                semantics = node["semantics"]
-                operator = semantics["operator"]
-                if operator == "invoke-operation":
-                    child = operations[instruction["operation"]["id"]]
-                    child_arguments: dict[str, dict[str, Any]] = {}
-                    for binding in instruction["arguments"]:
-                        operand = binding["operand"]
-                        if operand["kind"] == "port":
-                            actual = arguments[operand["port"]]
-                        elif operand["kind"] == "local":
-                            actual = locals_[operand["local"]]
-                        else:
-                            actual = {"value": operand["literal"]}
-                        child_arguments[binding["port"]] = actual
-                    try:
-                        child_outcome, child_result = execute(
-                            child,
-                            child_arguments,
-                            (*stack, selected["id"]),
-                            (*path, instruction["site"]),
-                        )
-                    except _ReferenceRuntimeRefusal as refusal:
-                        if resolved_call_sites is not None:
-                            refusal.call_site_identity = call_sites[
-                                (selected["id"], instruction["site"])
-                            ]["identity"]
-                        raise
-                    if resolved_call_sites is not None:
-                        call_site = call_sites[(selected["id"], instruction["site"])]
-                        outcome_row = next(
-                            row
-                            for row in call_site["outcomes"]
-                            if row["outcome"] == child_outcome
-                        )
-                        calls.append(
-                            {
-                                "call_site_identity": call_site["identity"],
-                                "site": "/".join((*path, instruction["site"])),
-                                "operation": call_site["operation"],
-                                "outcome": {
-                                    "id": child_outcome,
-                                    "identity": outcome_row["identity"],
-                                },
-                                "arguments": [
-                                    {
-                                        "formal_port_identity": row["port"]["identity"],
-                                        "actual_operand_identity": row["operand"][
-                                            "identity"
-                                        ],
-                                    }
-                                    for row in call_site["arguments"]
-                                ],
-                                "result_identity": call_site["result"]["identity"],
-                            }
-                        )
-                    result_binding = instruction["result"]
-                    if result_binding["kind"] == "local":
-                        write_local(result_binding["name"], child_result)
-                    elif result_binding["kind"] == "operation-result":
-                        operation_results[instruction["site"]] = child_result
-                    action = next(
-                        row["action"]
-                        for row in instruction["outcomes"]
-                        if row["outcome"] == child_outcome
-                    )
-                    if action["kind"] == "propagate":
-                        outcome = action["outcome"]
-                        break
-                    continue
-                if operator == "gameplay-precondition":
-                    if not _reference_compare(
-                        semantics["comparison"],
-                        integer(cell(instruction["left"])["value"]),
-                        integer(cell(instruction["right"])["value"]),
-                    ):
-                        outcome = instruction["outcome"]
-                        break
-                elif operator == "typed-require":
-                    if (
-                        cell(instruction["condition"])["value"]
-                        != instruction["expected"]
-                    ):
-                        refusal_reference = semantics["refusal_reference"]
-                        raise _ReferenceRuntimeRefusal(
-                            instruction[refusal_reference["instruction_member"]]
-                        )
-                elif operator == "guarded-outcome-block":
-                    if cell(instruction["condition"])["value"]:
-                        unit_contract = runtime["fixed_value_contracts"]["kernel-unit"]
-                        guarded = {
-                            "body": instruction["body"],
-                            "default_outcome": instruction["outcome"],
-                            "effects": selected["effects"],
-                            "id": selected["id"],
-                            "inputs": [],
-                            "outcomes": list(selected["outcomes"]),
-                            "refusals": selected["refusals"],
-                            "resource_bounds": selected["resource_bounds"],
-                            "result": {
-                                **unit_contract,
-                                "access": "read",
-                                "discardable": True,
-                                "id": "result",
-                                "source": {"kind": "unit"},
-                            },
-                        }
-                        execute(
-                            guarded,
-                            {**arguments, **locals_},
-                            stack,
-                            path,
-                        )
-                        outcome = instruction["outcome"]
-                        break
-                elif operator == "named-integer-draw":
-                    draw = _reference_rng_draw(
-                        runtime["named_rng"],
-                        seed,
-                        instruction["stream"],
-                        instruction["minimum"],
-                        instruction["maximum"],
-                        rng_states,
-                        rng_indices,
-                    )
-                    draws.append(draw)
-                    write_local(instruction["target"], draw["value"])
-                elif operator == "typed-literal":
-                    write_local(instruction["target"], instruction["literal"])
-                elif operator == "copy-value":
-                    write_local(
-                        instruction["target"],
-                        cell(instruction["value"])["value"],
-                    )
-                elif operator == "bounded-lookup":
-                    key = instruction["key"]
-                    container = cell(instruction["value"])["value"]
-                    selector = structured_law(container["type"], "bounded-lookup")[
-                        "selector"
-                    ]
-                    if selector == "local-index":
-                        key = locals_[key]["value"]
-                    write_local(
-                        instruction["target"],
-                        structured_lookup(container, key),
-                    )
-                elif operator == "canonical-equal":
-                    left = cell(instruction["left"])["value"]
-                    right = cell(instruction["right"])["value"]
-                    left_integer = project_integer(left)
-                    right_integer = project_integer(right)
-                    if isinstance(left, bool) and isinstance(right, bool):
-                        equal = left == right
-                    elif isinstance(left, bool) or isinstance(right, bool):
-                        raise AssertionError(
-                            "admitted equality operands used different representations"
-                        )
-                    elif left_integer is not None and right_integer is not None:
-                        if isinstance(left, dict) and isinstance(right, dict):
-                            assert left["type"] == right["type"]
-                        equal = left_integer == right_integer
-                    else:
-                        assert isinstance(left, dict) and isinstance(right, dict)
-                        assert left["type"] == right["type"]
-                        equal = left["value"] == right["value"]
-                    write_local(
-                        instruction["target"],
-                        equal,
-                    )
-                elif operator == "collection-is-empty":
-                    write_local(
-                        instruction["target"],
-                        structured_is_empty(cell(instruction["value"])["value"]),
-                    )
-                elif operator in {
-                    "integer-add",
-                    "integer-subtract",
-                    "integer-multiply",
-                    "integer-maximum",
-                }:
-                    left = integer(cell(instruction["left"])["value"])
-                    right = integer(cell(instruction["right"])["value"])
-                    result = {
-                        "integer-add": lambda: left + right,
-                        "integer-subtract": lambda: left - right,
-                        "integer-multiply": lambda: left * right,
-                        "integer-maximum": lambda: max(left, right),
-                    }[operator]()
-                    write_local(instruction["target"], exact(result))
-                elif operator == "integer-compare":
-                    write_local(
-                        instruction["target"],
-                        _reference_compare(
-                            semantics["comparison"],
-                            integer(cell(instruction["left"])["value"]),
-                            integer(cell(instruction["right"])["value"]),
-                        ),
-                    )
-                elif operator == "select-value":
-                    choice = (
-                        instruction["when_true"]
-                        if cell(instruction["condition"])["value"]
-                        else instruction["when_false"]
-                    )
-                    write_local(instruction["target"], cell(choice)["value"])
-                elif operator == "state-integer-subtract":
-                    target = arguments[instruction["symbol"]]
-                    target["value"] = exact(
-                        target["value"] - cell(instruction["value"])["value"],
-                        target,
-                    )
-                elif operator == "state-write":
-                    target = arguments[instruction["symbol"]]
-                    value = cell(instruction["value"])["value"]
-                    if (
-                        isinstance(target["value"], int)
-                        and not isinstance(target["value"], bool)
-                        and isinstance(value, dict)
-                        and set(value) == {"type", "value"}
-                    ):
-                        value = value["value"]
-                    target["value"] = (
-                        exact(value, target)
-                        if isinstance(value, int) and not isinstance(value, bool)
-                        else value
-                    )
-                else:
-                    raise AssertionError(
-                        f"unsupported operator in authority: {operator}"
-                    )
-        except _ReferenceRuntimeRefusal as refusal:
-            for key, value in snapshot.items():
-                frame_cells[key]["value"] = value
-            if refusal.operation is None:
-                refusal.operation = selected["id"]
-                refusal.call_path = "/".join(path)
-            raise
-
-        outcome_definition = next(
-            row for row in selected["outcomes"] if row["id"] == outcome
-        )
-        if outcome_definition["state_policy"] == "rollback":
-            for key, value in snapshot.items():
-                frame_cells[key]["value"] = value
-        source = selected["result"]["source"]
-        if outcome_definition["kind"] != "success":
-            result = None
-        elif source["kind"] == "local":
-            result = locals_[source["name"]]["value"]
-        elif source["kind"] == "port":
-            result = arguments[source["name"]]["value"]
-        elif source["kind"] == "operation-result":
-            result = operation_results[source["site"]]
-        else:
-            assert source["kind"] == "unit"
-            result = None
-        return outcome, result
-
-    if resolved_entrypoint is None:
-        root_arguments = [
-            {
-                "port": port["id"],
-                "operand": {"kind": "symbol", "symbol": port["id"]},
-            }
-            for port in operation["inputs"]
-        ]
-        root_frame = {
-            row["port"]: cells[row["operand"]["symbol"]] for row in root_arguments
-        }
-    else:
-        root_frame = {}
-        for argument in resolved_entrypoint["arguments"]:
-            operand = argument["operand"]
-            if operand["kind"] == "symbol":
-                symbol = operand["symbol"]
-                coordinate = (
-                    symbol["model"],
-                    symbol["module"],
-                    symbol["name"],
-                )
-                actual = cells[coordinate]
-            else:
-                actual = {"value": operand["value"]}
-            root_frame[argument["port"]["name"]] = actual
-    try:
-        outcome, result = execute(
-            operation,
-            root_frame,
-            path=((resolved_entrypoint["id"],) if resolved_entrypoint else ()),
-        )
-    except _ReferenceRuntimeRefusal as refusal:
-        return {
-            "refusal": {
-                "reason": refusal.reason,
-                "operation": refusal.operation,
-                "call_path": refusal.call_path,
-                "call_site_identity": refusal.call_site_identity,
-            },
-            "state_before": [
-                {"name": display_names[name], "value": before[name]}
-                for name in sorted(before)
-            ],
-            "state_after": [
-                {
-                    "name": display_names[name],
-                    "value": state_cells[name]["value"],
-                }
-                for name in sorted(state_cells)
-            ],
-        }
-    outcome_definition = next(
-        row for row in operation["outcomes"] if row["id"] == outcome
-    )
-    if (
-        resolved_entrypoint is not None
-        and resolved_entrypoint["result"]["kind"] == "symbol"
-        and outcome_definition["kind"] == "success"
-    ):
-        symbol = resolved_entrypoint["result"]["symbol"]
-        coordinate = (symbol["model"], symbol["module"], symbol["name"])
-        cells[coordinate] = {"value": result}
-        display_names[coordinate] = symbol["name"]
-    if outcome_definition["state_policy"] == "rollback":
-        for name, value in before.items():
-            state_cells[name]["value"] = value
-    facts: dict[str, Any] = {
-        display_names[name]: cell["value"] for name, cell in cells.items()
-    }
-    event = {
-        "operation": operation["id"],
-        "outcome": {
-            "id": outcome,
-            "kind": outcome_definition["kind"],
-        },
-        "facts": _reference_fact_rows(facts),
-        "state_before": [
-            {"name": display_names[name], "value": before[name]}
-            for name in sorted(before)
-        ],
-        "state_after": [
-            {
-                "name": display_names[name],
-                "value": state_cells[name]["value"],
-            }
-            for name in sorted(state_cells)
-        ],
-        "rng_draws": draws,
-        "schedules": [],
-        "cancellations": [],
-    }
-    if resolved_entrypoint is None:
-        event["result"] = result
-    if resolved_entrypoint is not None:
-        event["entrypoint"] = {
-            "id": resolved_entrypoint["id"],
-            "identity": resolved_entrypoint["identity"],
-        }
-        event["calls"] = calls
-    return event
 
 
 def _reference_evaluate_value_program_vector(
@@ -4115,6 +3338,45 @@ def test_terminal_audit_validation_rejects_coordinated_active_step_drift(
     assert audit["budget_counters"]["event_steps"] > 0
     assert experiment_evidence_module.validate_experiment_artifact_set(checked, values)
 
+    replay_rir = deepcopy(checked.rir)
+    entrypoint = next(
+        row
+        for row in replay_rir["entrypoints"]
+        if row["id"] == audit["refusing_event"]["event_spec"]["entrypoint"]
+    )
+    root_coordinate = operation_program_module.operation_coordinate(
+        entrypoint["operation"]
+    )
+    selected_semantics = replay_rir["selected_semantics"]
+    selected = next(
+        row
+        for row in selected_semantics["operations"]
+        if row["package"] == root_coordinate[0]
+        and row["definition"]["id"] == root_coordinate[2]
+    )
+    decoy = deepcopy(selected)
+    decoy["package"] = "example.decoy"
+    decoy["definition"]["body"] = []
+    selected_semantics["packages"].append({"id": "example.decoy", "version": "1.0.0"})
+    selected_semantics["operations"].append(decoy)
+    budget = audit["budget_counters"]
+    replay_profile = next(
+        row
+        for row in selected_semantics["runtime_profiles"]
+        if row["id"] == checked.value["runtime"]["profile"]
+    )
+    assert (
+        evidence_replay_module.attempted_operation_charge(
+            replace(checked, rir=replay_rir),
+            audit["refusing_event"],
+            audit["refusing_event"]["event_spec"],
+            node_steps_before_operation=budget["node_steps"] - budget["event_steps"],
+            bounds=replay_profile["resource_bounds"],
+            require_budget_breach=True,
+        )
+        == budget["event_steps"]
+    )
+
     drifted_audit = deepcopy(audit)
     drifted_audit["budget_counters"]["event_steps"] = 0
     drifted_audit["budget_counters"]["node_steps"] = drifted_audit[
@@ -4549,17 +3811,18 @@ def test_event_step_budget_resets_for_each_event(tmp_path, run_cli):
     checked = experiment_admission_module.check_experiment(str(specification_path))
     assert isinstance(checked, experiment_admission_module.CheckedExperiment)
     rir = deepcopy(checked.rir)
-    operations = {
-        row["definition"]["id"]: row["definition"]
-        for row in rir["selected_semantics"]["operations"]
-    }
+    operations = operation_program_module.selected_operation_index(
+        rir["selected_semantics"]
+    )
     entrypoint = next(row for row in rir["entrypoints"] if row["id"] == "combat.cast")
-    operation = operations[entrypoint["operation"]["id"]]
+    root_coordinate = operation_program_module.operation_coordinate(
+        entrypoint["operation"]
+    )
     runtime_nodes = experiment_runtime_module._runtime_nodes(checked)
     per_event_steps = sum(
         runtime_nodes[instruction["node"]]["resource_charge"]["amount"]
-        for instruction in experiment_admission_module._expanded_operation_body(
-            operation, operations
+        for instruction in operation_program_module.expanded_operation_body(
+            root_coordinate, operations
         )
     )
     runtime_profile = next(
@@ -4582,6 +3845,40 @@ def test_event_step_budget_resets_for_each_event(tmp_path, run_cli):
         )
         == 2
     )
+
+
+def test_runtime_selects_same_named_operation_by_exact_coordinate(tmp_path, run_cli):
+    specification_path = _write_built_experiment(tmp_path, run_cli)
+    checked = experiment_admission_module.check_experiment(str(specification_path))
+    assert isinstance(checked, experiment_admission_module.CheckedExperiment)
+    rir = deepcopy(checked.rir)
+    entrypoint = next(row for row in rir["entrypoints"] if row["id"] == "combat.cast")
+    root_coordinate = operation_program_module.operation_coordinate(
+        entrypoint["operation"]
+    )
+    selected_semantics = rir["selected_semantics"]
+    selected = next(
+        row
+        for row in selected_semantics["operations"]
+        if row["package"] == root_coordinate[0]
+        and row["definition"]["id"] == root_coordinate[2]
+    )
+    decoy = deepcopy(selected)
+    decoy["package"] = "example.decoy"
+    decoy["definition"]["default_outcome"] = "decoy-must-not-run"
+    selected_semantics["packages"].append({"id": "example.decoy", "version": "1.0.0"})
+    selected_semantics["operations"].append(decoy)
+
+    artifacts = experiment_runtime_module.evaluate_experiment(replace(checked, rir=rir))
+
+    assert isinstance(artifacts, experiment_runtime_module.EvaluationArtifacts)
+    runtime_event = next(
+        event
+        for event in artifacts.members["event-trace"].value["events"]
+        if event["operation"] is not None
+    )
+    assert runtime_event["operation"] == root_coordinate[2]
+    assert runtime_event["outcome"]["id"] != "decoy-must-not-run"
 
 
 def test_observation_formula_runs_once_after_same_time_transition_queue_drains(
@@ -5569,7 +4866,7 @@ def test_example_effective_accuracy_formula_exercises_its_minimum_clamp(
     seed = next(
         candidate
         for candidate in range(10_000)
-        if _reference_rng_draw(
+        if reference_rng_draw(
             runtime["named_rng"],
             candidate,
             "hit",
@@ -6997,7 +6294,7 @@ def test_kernel_runtime_contract_vectors_and_rng_execute_in_reference_evaluator(
         indices: dict[str, int] = {}
         draw = {}
         for _ in range(inputs["index"] + 1):
-            draw = _reference_rng_draw(
+            draw = reference_rng_draw(
                 runtime["named_rng"],
                 inputs["seed"],
                 inputs["stream"],
@@ -7034,7 +6331,7 @@ def test_reference_runtime_canonical_equality_compares_kernel_booleans():
         "result": {"source": {"kind": "local", "name": "same"}},
     }
 
-    event = _reference_execute_event(
+    event = reference_execute_event(
         kernel,
         operation,
         {operation["id"]: operation},
@@ -7081,7 +6378,7 @@ def test_package_operation_execution_vectors_preserve_integer_runtime_behavior()
             "id": vector["id"],
             "values": vector["input"]["values"],
         }
-        event = _reference_execute_event(
+        event = reference_execute_event(
             kernel,
             operation,
             operations,
@@ -7093,24 +6390,11 @@ def test_package_operation_execution_vectors_preserve_integer_runtime_behavior()
                 if row["access"] == "read-write"
             },
         )
-        projection = {
-            "completion": {"kind": "outcome", "id": event["outcome"]["id"]},
-            "result": (
-                {"kind": "value", "value": event["result"]}
-                if event["outcome"]["kind"] == "success"
-                else {"kind": "not-produced"}
-            ),
-            "rng_draws": [
-                {
-                    member: draw[member]
-                    for member in ("candidate_hex", "index", "stream", "value")
-                }
-                for draw in event["rng_draws"]
-            ],
-            "state_after": event["state_after"],
-        }
-        assert projection == vector["expect"]
-        replay = _reference_execute_event(
+        observations = operation_execution_observations(kernel, ldb, vector)
+        assert observations["production"] == observations["independent"]
+        assert observations["production"] == observations["expected"]
+        projection = observations["production"]
+        replay = reference_execute_event(
             kernel,
             operation,
             operations,
@@ -7139,127 +6423,11 @@ def test_package_operation_execution_vectors_preserve_integer_runtime_behavior()
     ]
 
 
-def _reference_operation_execution_projection(
-    kernel: dict[str, Any],
-    ldb: Any,
-    operations: dict[str, dict[str, Any]],
-    vector: dict[str, Any],
-) -> dict[str, Any]:
-    operation = operations[vector["operation"]]
-    event = _reference_execute_event(
-        kernel,
-        operation,
-        operations,
-        {"id": vector["id"], "values": vector["input"]["values"]},
-        seed=vector["input"]["seed"],
-        state_names={
-            row["id"] for row in operation["inputs"] if row["access"] == "read-write"
-        },
-        language_bundle=ldb,
-    )
-    state_after = {row["name"]: row["value"] for row in event["state_after"]}
-    state_names = [
-        row["id"] for row in operation["inputs"] if row["access"] == "read-write"
-    ]
-    if "refusal" in event:
-        return {
-            "completion": {
-                "kind": "refusal",
-                "reason": event["refusal"]["reason"],
-            },
-            "result": {"kind": "not-produced"},
-            "rng_draws": [],
-            "state_after": [
-                {"name": name, "value": state_after[name]} for name in state_names
-            ],
-        }
-    return {
-        "completion": {
-            "kind": "outcome",
-            "id": event["outcome"]["id"],
-        },
-        "result": (
-            {"kind": "value", "value": event["result"]}
-            if event["outcome"]["kind"] == "success"
-            else {"kind": "not-produced"}
-        ),
-        "rng_draws": [
-            {
-                member: draw[member]
-                for member in ("candidate_hex", "index", "stream", "value")
-            }
-            for draw in event["rng_draws"]
-        ],
-        "state_after": [
-            {"name": name, "value": state_after[name]} for name in state_names
-        ],
-    }
-
-
-def _operation_execution_vectors(ldb: Any) -> list[tuple[str, dict[str, Any]]]:
-    return [
-        (vector_set["package_id"], vector)
-        for vector_set in ldb.package_conformance_vector_sets
-        for vector in vector_set["vector_definitions"]
-        if vector.get("kind") == "operation-execution"
-    ]
-
-
-def _candidate_conformance_failures(
-    kernel: dict[str, Any],
-    ldb: Any,
-    *,
-    vector_overrides: dict[str, dict[str, Any]] | None = None,
-) -> list[dict[str, Any]]:
-    """Return bounded candidate-graph and execution-vector disagreements."""
-    production_admission = _consumer_a(kernel, ldb)
-    independent_admission = _consumer_b(kernel, ldb)
-    if production_admission != independent_admission:
-        return [
-            {
-                "kind": "admission-divergence",
-                "independent": independent_admission,
-                "production": production_admission,
-            }
-        ]
-    if production_admission["admitted"] is not True:
-        return [
-            {
-                "kind": "candidate-refused",
-                "observed": production_admission["diagnostics"],
-            }
-        ]
-    context = authority_module.admit_authority_context(kernel, ldb)
-    assert isinstance(context, authority_module.AdmittedAuthorityContext)
-    operations = {row["id"]: row for row in ldb["language"]["operations"]}
-    failures: list[dict[str, Any]] = []
-    overrides = vector_overrides or {}
-    for package_id, declared in _operation_execution_vectors(ldb):
-        vector = overrides.get(declared["id"], declared)
-        production = evaluate_operation_execution_vector(context, vector)
-        independent = _reference_operation_execution_projection(
-            kernel, ldb, operations, vector
-        )
-        if production != independent or production != vector["expect"]:
-            failures.append(
-                {
-                    "kind": "vector-divergence",
-                    "package": package_id,
-                    "vector": vector["id"],
-                    "expected": vector["expect"],
-                    "production": production,
-                    "independent": independent,
-                }
-            )
-    return failures
-
-
 def test_neutral_structured_operation_vectors_cover_control_paths():
     kernel, ldb = mutable_authorities()
-    operations = {row["id"]: row for row in ldb["language"]["operations"]}
     vectors = [
         vector
-        for package_id, vector in _operation_execution_vectors(ldb)
+        for package_id, _package_version, vector in operation_execution_vectors(ldb)
         if package_id == "standard.conformance.structured"
     ]
     assert {vector["id"] for vector in vectors} == {
@@ -7270,10 +6438,55 @@ def test_neutral_structured_operation_vectors_cover_control_paths():
     }
 
     for vector in vectors:
-        assert (
-            _reference_operation_execution_projection(kernel, ldb, operations, vector)
-            == vector["expect"]
+        observations = operation_execution_observations(kernel, ldb, vector)
+        assert observations["production"] == observations["independent"]
+        assert observations["production"] == observations["expected"]
+
+
+def test_operation_execution_conformance_adapters_remain_independent():
+    support_root = Path(__file__).parent
+
+    def imported_modules(name: str) -> set[str]:
+        tree = ast.parse(
+            (support_root / name).read_text(encoding="utf-8"), filename=name
         )
+        return {
+            node.module
+            for node in ast.walk(tree)
+            if isinstance(node, ast.ImportFrom) and node.module is not None
+        }
+
+    production = imported_modules("schema2_operation_execution_production_support.py")
+    independent = imported_modules("schema2_operation_execution_independent_support.py")
+
+    assert "schema2_operation_execution_independent_support" not in production
+    assert "schema2_operation_execution_production_support" not in independent
+    assert "gda_balancing.domain.runtime.execution" not in independent
+
+
+def test_operation_conformance_indexes_same_named_package_definitions_exactly():
+    language = {
+        "packages": [
+            {
+                "id": package_id,
+                "version": "1.0.0",
+                "semantic_closure": [
+                    {
+                        "authority_path": "language.operations",
+                        "definitions": [{"id": "shared", "marker": marker}],
+                    }
+                ],
+            }
+            for package_id, marker in (("example.a", "a"), ("example.b", "b"))
+        ]
+    }
+
+    conformance = conformance_operation_index({"language": language})
+    production = production_operation_index(language)
+
+    assert conformance[("example.a", "1.0.0", "shared")]["marker"] == "a"
+    assert conformance[("example.b", "1.0.0", "shared")]["marker"] == "b"
+    assert production == conformance
 
 
 def test_generation_operation_vectors_cover_success_fallback_and_refusals():
@@ -7281,7 +6494,7 @@ def test_generation_operation_vectors_cover_success_fallback_and_refusals():
 
     assert {
         vector["id"]: vector["category"]
-        for package_id, vector in _operation_execution_vectors(ldb)
+        for package_id, _package_version, vector in operation_execution_vectors(ldb)
         if package_id == "game.generation"
     } == {
         "generation.select.success": "deterministic-rng",
@@ -7303,7 +6516,7 @@ def test_mechanic_rollback_replay_vectors_repeat_without_state_or_rng_drift():
     }
     vectors = [
         vector
-        for package_id, vector in _operation_execution_vectors(ldb)
+        for package_id, _package_version, vector in operation_execution_vectors(ldb)
         if package_id in {"game.generation", "game.build"}
         and vector["category"] == "rollback-replay"
     ]
@@ -7327,12 +6540,20 @@ def test_mechanic_rollback_replay_vectors_repeat_without_state_or_rng_drift():
         assert vector["expect"]["rng_draws"] == []
 
         observed = [
-            evaluate_operation_execution_vector(context, vector),
-            evaluate_operation_execution_vector(context, vector),
-            _reference_operation_execution_projection(kernel, ldb, operations, vector),
-            _reference_operation_execution_projection(kernel, ldb, operations, vector),
+            operation_execution_observations(
+                kernel,
+                ldb,
+                vector,
+                context=context,
+            )
+            for _ in range(2)
         ]
-        assert all(result == vector["expect"] for result in observed), vector["id"]
+        assert all(
+            observation["production"]
+            == observation["independent"]
+            == observation["expected"]
+            for observation in observed
+        ), vector["id"]
 
 
 def test_generation_operation_owns_the_no_reward_discriminator():
@@ -7382,7 +6603,7 @@ def test_build_operation_vectors_cover_success_alternatives_and_refusal():
 
     assert {
         vector["id"]: vector["category"]
-        for package_id, vector in _operation_execution_vectors(ldb)
+        for package_id, _package_version, vector in operation_execution_vectors(ldb)
         if package_id == "game.build"
     } == {
         "build.replace.success": "positive",
@@ -7490,20 +6711,20 @@ def test_build_operation_guards_no_reward_before_plan_access_without_rng():
 
 def test_candidate_graph_executes_every_operation_vector_in_two_consumers():
     kernel, ldb = mutable_authorities()
-    assert _operation_execution_vectors(ldb)
+    assert operation_execution_vectors(ldb)
 
-    assert _candidate_conformance_failures(kernel, ldb) == []
+    assert candidate_conformance_failures(kernel, ldb) == []
 
 
 def test_candidate_graph_gate_identifies_an_operation_vector_divergence():
     kernel, ldb = mutable_authorities()
     package_id, vector = next(
         (package_id, deepcopy(candidate))
-        for package_id, candidate in _operation_execution_vectors(ldb)
+        for package_id, _package_version, candidate in operation_execution_vectors(ldb)
         if candidate["id"] == "structured.select.empty-outcome"
     )
     vector["expect"]["completion"]["id"] = "mutated-outcome"
-    failures = _candidate_conformance_failures(
+    failures = candidate_conformance_failures(
         kernel, ldb, vector_overrides={vector["id"]: vector}
     )
 
@@ -7513,6 +6734,33 @@ def test_candidate_graph_gate_identifies_an_operation_vector_divergence():
     assert failures[0]["vector"] == vector["id"]
     assert failures[0]["production"] == failures[0]["independent"]
     assert failures[0]["production"] != failures[0]["expected"]
+
+
+def test_candidate_graph_gate_identifies_an_adapter_divergence(monkeypatch):
+    kernel, ldb = mutable_authorities()
+    target = "structured.select.empty-outcome"
+    evaluate = operation_conformance_module.evaluate_operation_execution_vector
+
+    def divergent_production(context, vector, **owner):
+        observation = deepcopy(evaluate(context, vector, **owner))
+        if vector["id"] == target:
+            completion = cast(dict[str, Any], observation["completion"])
+            completion["id"] = "production-only-outcome"
+        return observation
+
+    monkeypatch.setattr(
+        operation_conformance_module,
+        "evaluate_operation_execution_vector",
+        divergent_production,
+    )
+
+    failures = candidate_conformance_failures(kernel, ldb)
+
+    assert len(failures) == 1
+    assert failures[0]["kind"] == "vector-divergence"
+    assert failures[0]["vector"] == target
+    assert failures[0]["production"] != failures[0]["independent"]
+    assert failures[0]["independent"] == failures[0]["expected"]
 
 
 def test_operation_execution_projection_preserves_declared_state_order():
@@ -7527,7 +6775,7 @@ def test_operation_execution_projection_preserves_declared_state_order():
     operation["inputs"] = [*read_only, *reversed(writable)]
     vectors = [
         vector
-        for _package_id, vector in _operation_execution_vectors(ldb)
+        for _package_id, _package_version, vector in operation_execution_vectors(ldb)
         if vector["operation"] == operation["id"]
     ]
     input_order = [row["id"] for row in operation["inputs"]]
@@ -8030,9 +7278,9 @@ def test_predispatch_capability_refusal_publishes_no_terminal_audit(
         base_damage=24,
     )
     monkeypatch.setattr(
-        experiment_runtime_module,
-        "_SUPPORTED_RUNTIME_OPERATORS",
-        experiment_runtime_module._SUPPORTED_RUNTIME_OPERATORS - {"integer-multiply"},
+        runtime_projection_module,
+        "SUPPORTED_RUNTIME_OPERATORS",
+        runtime_projection_module.SUPPORTED_RUNTIME_OPERATORS - {"integer-multiply"},
     )
     spec_path = tmp_path / "runtime-refusal-experiment.json"
     spec_path.write_text(json.dumps(specification), encoding="utf-8")
@@ -8212,30 +7460,31 @@ def test_evaluator_manifest_uses_selected_operation_closure_and_build_provenance
     checked = experiment_admission_module.check_experiment(str(specification))
     assert isinstance(checked, experiment_admission_module.CheckedExperiment)
 
-    first = experiment_runtime_module._evaluator_manifest(checked)
-    operations = {
-        row["definition"]["id"]: row["definition"]
-        for row in checked.rir["selected_semantics"]["operations"]
-    }
+    first = runtime_projection_module.evaluator_manifest(checked)
+    operations = operation_program_module.selected_operation_index(
+        checked.rir["selected_semantics"]
+    )
     entrypoints = {row["id"]: row for row in checked.rir["entrypoints"]}
     assert set(first.value["instruction_nodes"]) == {
         instruction["node"]
         for scenario in checked.value["scenarios"]
         for event in scenario["event_plan"]
-        for instruction in experiment_admission_module._expanded_operation_body(
-            operations[entrypoints[event["entrypoint"]]["operation"]["id"]],
+        for instruction in operation_program_module.expanded_operation_body(
+            operation_program_module.operation_coordinate(
+                entrypoints[event["entrypoint"]]["operation"]
+            ),
             operations,
         )
     }
     assert first.value["evaluator_build_identity"] == (
-        experiment_runtime_module._evaluator_build_identity()
+        runtime_projection_module.evaluator_build_identity()
     )
     monkeypatch.setattr(
-        experiment_runtime_module,
-        "_evaluator_build_identity",
+        runtime_projection_module,
+        "evaluator_build_identity",
         lambda: "sha256:" + ("0" * 64),
     )
-    changed_build = experiment_runtime_module._evaluator_manifest(checked)
+    changed_build = runtime_projection_module.evaluator_manifest(checked)
     assert (
         changed_build.value["implementation_identity"]
         != (first.value["implementation_identity"])
@@ -8244,9 +7493,13 @@ def test_evaluator_manifest_uses_selected_operation_closure_and_build_provenance
 
 
 def test_operation_closure_includes_guard_body_nodes_and_invocations():
+    root = ("example", "1.0.0", "example.root")
+    child = ("example", "1.0.0", "example.child")
     operations = {
-        "example.root": {
+        root: {
             "id": "example.root",
+            "effects": ["event.commit"],
+            "refusals": ["example.root-refusal"],
             "body": [
                 {
                     "node": "guard-block",
@@ -8260,21 +7513,22 @@ def test_operation_closure_includes_guard_body_nodes_and_invocations():
                                 "version": "1.0.0",
                                 "id": "example.child",
                             },
+                            "site": "child",
                         },
                     ],
                     "outcome": "complete",
                 }
             ],
         },
-        "example.child": {
+        child: {
             "id": "example.child",
+            "effects": ["snapshot.commit"],
+            "refusals": ["example.child-refusal"],
             "body": [{"node": "constant", "literal": 1, "target": "one"}],
         },
     }
 
-    expanded = experiment_admission_module._expanded_operation_body(
-        operations["example.root"], operations
-    )
+    expanded = operation_program_module.expanded_operation_body(root, operations)
 
     assert [instruction["node"] for instruction in expanded] == [
         "guard-block",
@@ -8282,6 +7536,139 @@ def test_operation_closure_includes_guard_body_nodes_and_invocations():
         "invoke",
         "constant",
     ]
+    projection = operation_program_module.project_operation_program(
+        root,
+        operations,
+        operation_node_ids={"invoke"},
+        invocation_node_ids={"invoke"},
+    )
+    assert projection.reachable_operations == {root, child}
+    assert projection.invocation_paths == ((("child",), child),)
+    assert projection.node_ids == {"guard-block", "copy", "invoke", "constant"}
+    assert projection.effects == {"event.commit", "snapshot.commit"}
+    assert projection.refusals == {
+        "example.child-refusal",
+        "example.root-refusal",
+    }
+    assert projection.resource_charge == 4
+
+
+def test_operation_program_resolves_same_named_operations_by_exact_coordinate():
+    root = ("example.root", "1.0.0", "example.root")
+    wrong = ("example.a", "1.0.0", "shared")
+    selected = ("example.b", "1.0.0", "shared")
+    operations = {
+        root: {
+            "id": "example.root",
+            "body": [
+                {
+                    "node": "invoke",
+                    "operation": {
+                        "package": selected[0],
+                        "version": selected[1],
+                        "id": selected[2],
+                    },
+                    "site": "selected",
+                }
+            ],
+        },
+        wrong: {"id": "shared", "body": [{"node": "wrong"}]},
+        selected: {"id": "shared", "body": [{"node": "selected"}]},
+    }
+
+    expanded = operation_program_module.expanded_operation_body(root, operations)
+
+    assert [instruction["node"] for instruction in expanded] == [
+        "invoke",
+        "selected",
+    ]
+
+
+def test_operation_program_projects_descendants_for_each_invocation_path():
+    root = ("example", "1.0.0", "root")
+    child = ("example", "1.0.0", "child")
+    grandchild = ("example", "1.0.0", "grandchild")
+
+    def reference(coordinate):
+        return {
+            "package": coordinate[0],
+            "version": coordinate[1],
+            "id": coordinate[2],
+        }
+
+    operations = {
+        root: {
+            "id": "root",
+            "effects": [],
+            "refusals": [],
+            "body": [
+                {"node": "invoke", "operation": reference(child), "site": "a"},
+                {"node": "invoke", "operation": reference(child), "site": "b"},
+            ],
+        },
+        child: {
+            "id": "child",
+            "effects": [],
+            "refusals": [],
+            "body": [
+                {
+                    "node": "invoke",
+                    "operation": reference(grandchild),
+                    "site": "g",
+                }
+            ],
+        },
+        grandchild: {
+            "id": "grandchild",
+            "effects": [],
+            "refusals": [],
+            "body": [],
+        },
+    }
+
+    projection = operation_program_module.project_operation_program(
+        root,
+        operations,
+        operation_node_ids={"invoke"},
+        invocation_node_ids={"invoke"},
+    )
+
+    assert projection.invocation_paths == (
+        (("a",), child),
+        (("a", "g"), grandchild),
+        (("b",), child),
+        (("b", "g"), grandchild),
+    )
+
+
+def test_operation_program_projects_guard_expanded_audit_positions():
+    body = [
+        {"node": "constant"},
+        {
+            "node": "guard-block",
+            "body": [{"node": "copy"}, {"node": "require"}],
+        },
+        {"node": "copy"},
+    ]
+
+    indices = operation_program_module.guard_expanded_instruction_indices(body)
+
+    assert indices == (0, 1, 4)
+
+
+def test_operation_program_owns_formula_instruction_provenance():
+    operation: dict[str, Any] = {"extensions": {}}
+    operation_program_module.record_instruction_evaluation_sites(
+        operation,
+        first_instruction_index=2,
+        instruction_count=2,
+        evaluation_site_identity="sha256:" + "a" * 64,
+    )
+
+    assert operation_program_module.instruction_evaluation_sites(operation) == {
+        2: "sha256:" + "a" * 64,
+        3: "sha256:" + "a" * 64,
+    }
 
 
 def test_metric_dataset_carries_the_complete_bounded_metric_contract(tmp_path, run_cli):
@@ -8546,7 +7933,7 @@ def _assert_numeric_overflow_rolls_back_the_entire_current_event(
     resolved_entrypoint = next(
         row for row in rir["entrypoints"] if row["id"] == "combat.cast"
     )
-    reference = _reference_execute_event(
+    reference = reference_execute_event(
         kernel,
         operations["game.combat.cast-v1"],
         operations,
@@ -8739,7 +8126,7 @@ def test_ordered_writable_aliases_share_one_runtime_location(tmp_path, run_cli):
     assert isinstance(production, experiment_runtime_module.EvaluationArtifacts)
     production_event = production.members["event-trace"].value["events"][0]
     entrypoint = next(row for row in rir["entrypoints"] if row["id"] == "combat.cast")
-    reference_event = _reference_execute_event(
+    reference_event = reference_execute_event(
         checked.kernel,
         operations["game.combat.cast-v1"],
         operations,
@@ -8801,7 +8188,7 @@ def test_nested_integer_literal_is_observable_across_evaluators(tmp_path, run_cl
     assert isinstance(production, experiment_runtime_module.EvaluationArtifacts)
     production_event = production.members["event-trace"].value["events"][0]
     entrypoint = next(row for row in rir["entrypoints"] if row["id"] == "combat.cast")
-    reference_event = _reference_execute_event(
+    reference_event = reference_execute_event(
         checked.kernel,
         cast_operation,
         operations,
@@ -8866,7 +8253,7 @@ def test_nested_operation_result_is_observable_across_evaluators(tmp_path, run_c
     assert isinstance(production, experiment_runtime_module.EvaluationArtifacts)
     production_event = production.members["event-trace"].value["events"][0]
     entrypoint = next(row for row in rir["entrypoints"] if row["id"] == "combat.cast")
-    reference_event = _reference_execute_event(
+    reference_event = reference_execute_event(
         checked.kernel,
         cast_operation,
         operations,
@@ -8967,7 +8354,7 @@ def test_ordered_writable_alias_write_is_visible_to_later_child_call(
 
     assert isinstance(production, experiment_runtime_module.EvaluationArtifacts)
     production_event = production.members["event-trace"].value["events"][0]
-    reference_event = _reference_execute_event(
+    reference_event = reference_execute_event(
         checked.kernel,
         cast_operation,
         operations,
